refactor(mcp): decompose src/mcp.rs into transport/schema/audit/tools modules (Issue #344)
Splits the 3324-line mcp.rs into a directory module:
- mcp/mod.rs: JSON-RPC transport, McpServer struct, request dispatch, stdio loop
- mcp/schema.rs: tools/list tool definitions (previously an inline 318-line
  json! block in a single method)
- mcp/audit.rs: audit logging, sandbox glue, ruff-tool-runner helpers
- mcp/tools/: per-tool handlers grouped by concern (resolve.rs, install.rs,
  run.rs, maintenance.rs for gc/doctor/lint/type_check/profile/fix/context/
  drift/test/audit/upgrade)

All 14 call_*/read_* tool methods never referenced `self`, so they were
converted to free functions instead of requiring struct-field visibility
workarounds. This deviates from the issue's literal proposed layout (no
separate rpc.rs; JSON-RPC framing stays in mod.rs alongside McpServer) since
splitting the transport out further would leave two tightly-coupled halves
with no clear seam.

Zero behavior changes: byte-identical logic, mechanically moved and
regrouped. All existing MCP unit/integration tests pass unchanged (68 mcp.rs
tests; 4 pre-existing failures are a local dyld/venv environment issue
unrelated to this change, reproduced identically on main).

Progress on #344 — closes the schema-inlining and giant-impl-block pain
points; not claiming full literal-layout closure.
