--- src/mcp.rs
@@ -1,3324 +1,0 @@
-//! MCP (Model Context Protocol) Server Implementation
-//!
-//! PR4.3: MCP server for programmatic control of PyBun.
-//! PR4.3b: Implemented real tool execution (resolve, install, run, doctor).
-//!
-//! This module implements the MCP protocol for AI agents and tools to
-//! interact with PyBun via JSON-RPC.
-//!
-//! ## Supported Methods
-//! - `initialize`: Initialize the MCP session
-//! - `tools/list`: List available tools
-//! - `tools/call`: Call a tool
-//! - `resources/list`: List available resources
-//! - `shutdown`: Shutdown the server
-//!
-//! ## Tools
-//! - `pybun_resolve`: Resolve dependencies
-//! - `pybun_install`: Install packages
-//! - `pybun_run`: Run Python scripts
-//! - `pybun_gc`: Run garbage collection
-//! - `pybun_doctor`: Run environment diagnostics
-
-use serde::{Deserialize, Serialize};
-use serde_json::{Value, json};
-use std::collections::{BTreeMap, VecDeque};
-use std::fs::{self, OpenOptions};
-use std::io::Write as IoWrite;
-use std::path::{Path, PathBuf};
-use std::process::Command as ProcessCommand;
-use std::time::{Instant, SystemTime, UNIX_EPOCH};
-use tokio::io::{AsyncBufReadExt, AsyncWriteExt, BufReader};
-use uuid::Uuid;
-
-/// MCP Protocol version we support
-pub const PROTOCOL_VERSION: &str = "2024-11-05";
-
-/// Server name and version
-pub const SERVER_NAME: &str = "pybun-mcp";
-pub const SERVER_VERSION: &str = env!("CARGO_PKG_VERSION");
-
-/// JSON-RPC request structure
-#[derive(Debug, Deserialize)]
-pub struct JsonRpcRequest {
-    pub jsonrpc: String,
-    pub method: String,
-    #[serde(default)]
-    pub params: Value,
-    pub id: Option<Value>,
-}
-
-/// JSON-RPC response structure
-#[derive(Debug, Serialize)]
-pub struct JsonRpcResponse {
-    pub jsonrpc: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result: Option<Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<JsonRpcError>,
-    pub id: Value,
-}
-
-/// JSON-RPC error structure
-#[derive(Debug, Serialize)]
-pub struct JsonRpcError {
-    pub code: i32,
-    pub message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Value>,
-}
-
-/// MCP Tool definition
-#[derive(Debug, Serialize)]
-pub struct Tool {
-    pub name: String,
-    pub description: String,
-    #[serde(rename = "inputSchema")]
-    pub input_schema: Value,
-}
-
-/// MCP Resource definition
-#[derive(Debug, Serialize)]
-pub struct Resource {
-    pub uri: String,
-    pub name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(rename = "mimeType", skip_serializing_if = "Option::is_none")]
-    pub mime_type: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-struct AuditConfig {
-    enabled: bool,
-    path: Option<PathBuf>,
-    hash_inputs: bool,
-    retention_days: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct AuditToolSchema {
-    name: String,
-    schema_version: u64,
-    protocol_version: &'static str,
-    server_version: &'static str,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct AuditOutputSummary {
-    status: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    exit_code: Option<i64>,
-    stdout_bytes: u64,
-    stderr_bytes: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct AuditSandboxEvent {
-    #[serde(rename = "type")]
-    event_type: String,
-    count: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct AuditFileWrite {
-    path: String,
-    size_bytes: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    sha256_before: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    sha256_after: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct AuditEntry {
-    session_id: String,
-    call_id: String,
-    timestamp: String,
-    tool: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    input: Option<Value>,
-    input_hash: String,
-    tool_schema: AuditToolSchema,
-    output_summary: AuditOutputSummary,
-    sandbox_events: Vec<AuditSandboxEvent>,
-    file_writes: Vec<AuditFileWrite>,
-    duration_ms: u64,
-}
-
-#[derive(Debug, Clone)]
-struct FileState {
-    size_bytes: u64,
-    sha256: String,
-}
-
-type FileSnapshot = BTreeMap<PathBuf, FileState>;
-
-#[derive(Debug)]
-struct McpAuditLog {
-    config: AuditConfig,
-    recent: VecDeque<AuditEntry>,
-}
-
-#[derive(Debug, Clone)]
-struct ToolRunner {
-    program: PathBuf,
-    base_args: Vec<String>,
-}
-
-impl ToolRunner {
-    fn new(program: impl Into<PathBuf>) -> Self {
-        Self {
-            program: program.into(),
-            base_args: Vec::new(),
-        }
-    }
-
-    fn python_module(python_path: impl Into<PathBuf>, module: &str) -> Self {
-        Self {
-            program: python_path.into(),
-            base_args: vec!["-m".to_string(), module.to_string()],
-        }
-    }
-
-    fn command(&self) -> ProcessCommand {
-        let mut cmd = ProcessCommand::new(&self.program);
-        cmd.args(&self.base_args);
-        cmd
-    }
-}
-
-fn current_working_dir() -> Result<PathBuf, String> {
-    std::env::current_dir().map_err(|e| e.to_string())
-}
-
-/// Whether git was available when changed-file detection ran.
-#[derive(Debug)]
-enum GitAvailability {
-    /// git ran successfully (working dir is a repository)
-    Available,
-    /// git is not installed or this is not a git repository
-    NotARepo,
-}
-
-/// Run a git command with a 5-second timeout and return its stdout lines.
-/// Returns `Err` only on spawn failure; a non-zero exit code or timeout yields `Ok(None)`.
-fn run_git_with_timeout(args: &[&str], working_dir: &Path) -> Option<String> {
-    use std::time::Duration;
-
-    let mut child = std::process::Command::new("git")
-        .args(args)
-        .current_dir(working_dir)
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::null())
-        .spawn()
-        .ok()?;
-
-    let deadline = std::time::Instant::now() + Duration::from_secs(5);
-    loop {
-        match child.try_wait().ok()? {
-            Some(status) if status.success() => {
-                return child
-                    .wait_with_output()
-                    .ok()
-                    .map(|o| String::from_utf8_lossy(&o.stdout).into_owned());
-            }
-            Some(_) => return None,
-            None => {
-                if std::time::Instant::now() >= deadline {
-                    let _ = child.kill();
-                    return None;
-                }
-                std::thread::sleep(std::time::Duration::from_millis(50));
-            }
-        }
-    }
-}
-
-/// Detect Python test files changed since last git commit (modified + new untracked).
-/// Source files are mapped to test files via naming convention (src/foo.py → tests/test_foo.py).
-/// Returns `(paths, NotARepo)` when git is unavailable or the directory is not a repository.
-fn get_changed_test_files(working_dir: &Path) -> Result<(Vec<PathBuf>, GitAvailability), String> {
-    use std::collections::HashSet;
-
-    let mut seen: HashSet<PathBuf> = HashSet::new();
-    let mut changed_py_files: Vec<PathBuf> = Vec::new();
-    let mut git_ran = false;
-
-    // Modified and staged files vs HEAD
-    if let Some(stdout) = run_git_with_timeout(&["diff", "--name-only", "HEAD"], working_dir) {
-        git_ran = true;
-        for line in stdout.lines() {
-            if line.ends_with(".py") {
-                let path = working_dir.join(line);
-                if seen.insert(path.clone()) {
-                    changed_py_files.push(path);
-                }
-            }
-        }
-    }
-
-    // Untracked new files not yet committed
-    if let Some(stdout) =
-        run_git_with_timeout(&["ls-files", "--others", "--exclude-standard"], working_dir)
-    {
-        git_ran = true;
-        for line in stdout.lines() {
-            if line.ends_with(".py") {
-                let path = working_dir.join(line);
-                if seen.insert(path.clone()) {
-                    changed_py_files.push(path);
-                }
-            }
-        }
-    }
-
-    if !git_ran {
-        return Ok((Vec::new(), GitAvailability::NotARepo));
-    }
-
-    let mut test_seen: HashSet<PathBuf> = HashSet::new();
-    let mut test_paths: Vec<PathBuf> = Vec::new();
-
-    for file in changed_py_files {
-        let file_name = file.file_name().and_then(|n| n.to_str()).unwrap_or("");
-        if file_name.starts_with("test_") || file_name.ends_with("_test.py") {
-            if file.exists() && test_seen.insert(file.clone()) {
-                test_paths.push(file);
-            }
-        } else if file_name.ends_with(".py") {
-            // Convention-based mapping: src/foo.py → tests/test_foo.py
-            let stem = file.file_stem().and_then(|s| s.to_str()).unwrap_or("");
-            let test_name = format!("test_{}.py", stem);
-            let candidates: Vec<PathBuf> = [
-                Some(working_dir.join("tests").join(&test_name)),
-                Some(working_dir.join(&test_name)),
-                file.parent()
-                    .and_then(|p| p.parent())
-                    .map(|pp| pp.join("tests").join(&test_name)),
-            ]
-            .into_iter()
-            .flatten()
-            .collect();
-
-            for candidate in candidates {
-                if candidate.exists() && test_seen.insert(candidate.clone()) {
-                    test_paths.push(candidate);
-                    break;
-                }
-            }
-        }
-    }
-
-    Ok((test_paths, GitAvailability::Available))
-}
-
-fn valid_ruff_status(status: &std::process::ExitStatus) -> bool {
-    matches!(status.code(), Some(0 | 1))
-}
-
-fn ruff_failure_message(action: &str, stdout: &str, stderr: &str) -> String {
-    let detail = if stderr.trim().is_empty() {
-        stdout.trim()
-    } else {
-        stderr.trim()
-    };
-    if detail.is_empty() {
-        format!("ruff {action} failed")
-    } else {
-        format!("ruff {action} failed: {detail}")
-    }
-}
-
-fn parse_ruff_violations(stdout: &str) -> Result<Vec<Value>, String> {
-    if stdout.trim().is_empty() {
-        return Ok(vec![]);
-    }
-
-    serde_json::from_str::<Vec<Value>>(stdout)
-        .map_err(|e| format!("Failed to parse ruff JSON output: {e}"))
-}
-
-fn find_env_executable(python_path: &Path, tool_name: &str) -> Option<PathBuf> {
-    let bin_dir = python_path.parent()?;
-    let candidates: Vec<String> = if cfg!(windows) {
-        vec![
-            format!("{tool_name}.exe"),
-            format!("{tool_name}.cmd"),
-            format!("{tool_name}.bat"),
-            tool_name.to_string(),
-        ]
-    } else {
-        vec![tool_name.to_string()]
-    };
-
-    candidates
-        .iter()
-        .map(|candidate| bin_dir.join(candidate))
-        .find(|candidate| candidate.is_file())
-}
-
-fn resolve_ruff_runner(working_dir: &Path) -> Result<Option<ToolRunner>, String> {
-    use crate::env::find_python_env;
-
-    if let Ok(env) = find_python_env(working_dir) {
-        if let Some(ruff_path) = find_env_executable(&env.python_path, "ruff") {
-            return Ok(Some(ToolRunner::new(ruff_path)));
-        }
-
-        let module_runner = ToolRunner::python_module(env.python_path, "ruff");
-        if module_runner
-            .command()
-            .arg("--version")
-            .output()
-            .map(|output| output.status.success())
-            .unwrap_or(false)
-        {
-            return Ok(Some(module_runner));
-        }
-    }
-
-    let global_runner = ToolRunner::new("ruff");
-    if global_runner
-        .command()
-        .arg("--version")
-        .output()
-        .map(|output| output.status.success())
-        .unwrap_or(false)
-    {
-        return Ok(Some(global_runner));
-    }
-
-    Ok(None)
-}
-
-impl AuditConfig {
-    fn load() -> Self {
-        let cwd = std::env::current_dir().unwrap_or_else(|_| PathBuf::from("."));
-        let (project_root, mut config) = load_pyproject_audit_config(&cwd).unwrap_or_else(|| {
-            (
-                cwd.clone(),
-                Self {
-                    enabled: true,
-                    path: default_audit_log_path(),
-                    hash_inputs: false,
-                    retention_days: 30,
-                },
-            )
-        });
-
-        if config.path.is_none() && config.enabled {
-            config.path = default_audit_log_path();
-        }
-
-        if let Some(path) = config.path.take() {
-            config.path = Some(resolve_audit_path(path, &project_root));
-        }
-
-        if let Ok(path) = std::env::var("PYBUN_AUDIT_LOG") {
-            if path == "/dev/null" {
-                config.enabled = false;
-                config.path = None;
-            } else if !path.trim().is_empty() {
-                config.enabled = true;
-                config.path = Some(resolve_audit_path(PathBuf::from(path), &cwd));
-            }
-        }
-
-        config
-    }
-}
-
-impl McpAuditLog {
-    fn new() -> Self {
-        Self {
-            config: AuditConfig::load(),
-            recent: VecDeque::new(),
-        }
-    }
-
-    fn prepare_for_call(&self, tool_args: &Value) -> Result<(), String> {
-        if !self.config.enabled {
-            return Ok(());
-        }
-
-        let Some(path) = &self.config.path else {
-            return Ok(());
-        };
-
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent).map_err(|e| {
-                format!(
-                    "failed to create audit log directory {}: {e}",
-                    parent.display()
-                )
-            })?;
-        }
-
-        OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(path)
-            .map_err(|e| format!("failed to create audit log {}: {e}", path.display()))?;
-
-        if audit_path_conflicts_with_allow_write(path, tool_args) {
-            return Err(format!(
-                "audit log path {} must be outside sandbox allow_write paths",
-                path.display()
-            ));
-        }
-
-        Ok(())
-    }
-
-    fn record(&mut self, entry: AuditEntry) {
-        if !self.config.enabled {
-            return;
-        }
-
-        self.recent.push_back(entry.clone());
-        while self.recent.len() > 20 {
-            self.recent.pop_front();
-        }
-
-        let Some(path) = &self.config.path else {
-            return;
-        };
-        let write_result = (|| -> Result<(), String> {
-            let mut file = OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(path)
-                .map_err(|e| e.to_string())?;
-            let line = serde_json::to_string(&entry).map_err(|e| e.to_string())?;
-            writeln!(file, "{line}").map_err(|e| e.to_string())
-        })();
-
-        if let Err(err) = write_result {
-            eprintln!("warning: failed to append MCP audit log: {err}");
-        }
-    }
-
-    fn recent_json(&self, session_id: &str) -> String {
-        json!({
-            "session_id": session_id,
-            "count": self.recent.len(),
-            "entries": self.recent,
-            "retention_days": self.config.retention_days,
-        })
-        .to_string()
-    }
-}
-
-fn default_audit_log_path() -> Option<PathBuf> {
-    crate::cache::Cache::new().ok().map(|cache| {
-        cache
-            .logs_dir()
-            .join(format!("mcp-audit-{}.jsonl", utc_date_now()))
-    })
-}
-
-fn load_pyproject_audit_config(start: &Path) -> Option<(PathBuf, AuditConfig)> {
-    let mut current = start.to_path_buf();
-    loop {
-        let candidate = current.join("pyproject.toml");
-        if candidate.exists() {
-            let content = fs::read_to_string(&candidate).ok()?;
-            let value: toml::Value = toml::from_str(&content).ok()?;
-            let audit = value.get("tool")?.get("pybun")?.get("mcp")?.get("audit")?;
-
-            let enabled = audit
-                .get("enabled")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(true);
-            let path = audit
-                .get("path")
-                .and_then(|v| v.as_str())
-                .map(PathBuf::from);
-            let hash_inputs = audit
-                .get("hash_inputs")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-            let retention_days = audit
-                .get("retention_days")
-                .and_then(|v| v.as_integer())
-                .and_then(|v| u64::try_from(v).ok())
-                .unwrap_or(30);
-
-            return Some((
-                current,
-                AuditConfig {
-                    enabled,
-                    path,
-                    hash_inputs,
-                    retention_days,
-                },
-            ));
-        }
-
-        if !current.pop() {
-            return None;
-        }
-    }
-}
-
-fn resolve_audit_path(path: PathBuf, base: &Path) -> PathBuf {
-    let path_string = path.to_string_lossy();
-    if path_string == "~" {
-        return dirs::home_dir().unwrap_or_else(|| base.to_path_buf());
-    }
-    if let Some(stripped) = path_string.strip_prefix("~/") {
-        return dirs::home_dir()
-            .unwrap_or_else(|| base.to_path_buf())
-            .join(stripped);
-    }
-    if path.is_absolute() {
-        path
-    } else {
-        base.join(path)
-    }
-}
-
-fn audit_path_conflicts_with_allow_write(audit_path: &Path, tool_args: &Value) -> bool {
-    let Some(allow_write) = tool_args
-        .get("sandbox_policy")
-        .and_then(|p| p.get("allow_write"))
-        .and_then(|v| v.as_array())
-    else {
-        return false;
-    };
-
-    let audit_path = normalized_absolute_path(audit_path);
-    allow_write.iter().filter_map(|v| v.as_str()).any(|path| {
-        let allowed = normalized_absolute_path(Path::new(path));
-        path_within(&audit_path, &allowed)
-    })
-}
-
-fn normalized_absolute_path(path: &Path) -> PathBuf {
-    let absolute = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        std::env::current_dir()
-            .unwrap_or_else(|_| PathBuf::from("."))
-            .join(path)
-    };
-    absolute.canonicalize().unwrap_or(absolute)
-}
-
-fn path_within(path: &Path, root: &Path) -> bool {
-    path == root || path.starts_with(root)
-}
-
-fn utc_date_now() -> String {
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default();
-    let days = (now.as_secs() / 86_400) as i64;
-    let (year, month, day) = civil_from_unix_days(days);
-    format!("{year:04}-{month:02}-{day:02}")
-}
-
-fn utc_timestamp_now() -> String {
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default();
-    let total_secs = now.as_secs();
-    let days = (total_secs / 86_400) as i64;
-    let secs_of_day = total_secs % 86_400;
-    let (year, month, day) = civil_from_unix_days(days);
-    let hour = secs_of_day / 3_600;
-    let minute = (secs_of_day % 3_600) / 60;
-    let second = secs_of_day % 60;
-    let micros = now.subsec_micros();
-    format!("{year:04}-{month:02}-{day:02}T{hour:02}:{minute:02}:{second:02}.{micros:06}Z")
-}
-
-fn civil_from_unix_days(days_since_epoch: i64) -> (i32, u32, u32) {
-    let z = days_since_epoch + 719_468;
-    let era = if z >= 0 { z } else { z - 146_096 } / 146_097;
-    let doe = z - era * 146_097;
-    let yoe = (doe - doe / 1_460 + doe / 36_524 - doe / 146_096) / 365;
-    let y = yoe + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let day = doy - (153 * mp + 2) / 5 + 1;
-    let month = mp + if mp < 10 { 3 } else { -9 };
-    let year = y + if month <= 2 { 1 } else { 0 };
-    (year as i32, month as u32, day as u32)
-}
-
-fn input_hash(input: &Value) -> String {
-    let bytes = serde_json::to_vec(input).unwrap_or_default();
-    format!("sha256:{}", crate::security::sha256_bytes(&bytes))
-}
-
-fn build_audit_entry(
-    session_id: &str,
-    tool_name: &str,
-    tool_args: &Value,
-    result: &Result<String, String>,
-    hash_inputs: bool,
-    file_writes: Vec<AuditFileWrite>,
-    duration_ms: u64,
-) -> AuditEntry {
-    AuditEntry {
-        session_id: session_id.to_string(),
-        call_id: Uuid::new_v4().to_string(),
-        timestamp: utc_timestamp_now(),
-        tool: tool_name.to_string(),
-        input: (!hash_inputs).then(|| tool_args.clone()),
-        input_hash: input_hash(tool_args),
-        tool_schema: AuditToolSchema {
-            name: tool_name.to_string(),
-            schema_version: 1,
-            protocol_version: PROTOCOL_VERSION,
-            server_version: SERVER_VERSION,
-        },
-        output_summary: summarize_tool_result(result),
-        sandbox_events: sandbox_events_from_result(result),
-        file_writes,
-        duration_ms,
-    }
-}
-
-fn summarize_tool_result(result: &Result<String, String>) -> AuditOutputSummary {
-    match result {
-        Ok(content) => {
-            let parsed = serde_json::from_str::<Value>(content).ok();
-            let status = parsed
-                .as_ref()
-                .and_then(|value| value.get("status"))
-                .and_then(|value| value.as_str())
-                .map(|status| {
-                    if status == "error" {
-                        "error".to_string()
-                    } else {
-                        "ok".to_string()
-                    }
-                })
-                .unwrap_or_else(|| "ok".to_string());
-            let exit_code = parsed
-                .as_ref()
-                .and_then(|value| value.get("exit_code"))
-                .and_then(|value| value.as_i64());
-            let stdout_bytes = parsed
-                .as_ref()
-                .and_then(|value| value.get("stdout"))
-                .and_then(|value| value.as_str())
-                .map(|stdout| stdout.len() as u64)
-                .unwrap_or(0);
-            let stderr_bytes = parsed
-                .as_ref()
-                .and_then(|value| value.get("stderr"))
-                .and_then(|value| value.as_str())
-                .map(|stderr| stderr.len() as u64)
-                .unwrap_or(0);
-            AuditOutputSummary {
-                status,
-                exit_code,
-                stdout_bytes,
-                stderr_bytes,
-            }
-        }
-        Err(message) => AuditOutputSummary {
-            status: "error".to_string(),
-            exit_code: None,
-            stdout_bytes: 0,
-            stderr_bytes: message.len() as u64,
-        },
-    }
-}
-
-fn sandbox_events_from_result(result: &Result<String, String>) -> Vec<AuditSandboxEvent> {
-    let Ok(content) = result else {
-        return vec![];
-    };
-    let Ok(value) = serde_json::from_str::<Value>(content) else {
-        return vec![];
-    };
-    let Some(audit) = value.get("audit").and_then(|v| v.as_object()) else {
-        return vec![];
-    };
-
-    [
-        ("blocked_subprocesses", "blocked_subprocess"),
-        ("blocked_network", "blocked_network"),
-        ("blocked_file_reads", "blocked_file_read"),
-        ("blocked_file_writes", "blocked_file_write"),
-    ]
-    .iter()
-    .filter_map(|(key, event_type)| {
-        let count = audit.get(*key).and_then(|v| v.as_u64()).unwrap_or(0);
-        (count > 0).then(|| AuditSandboxEvent {
-            event_type: (*event_type).to_string(),
-            count,
-        })
-    })
-    .collect()
-}
-
-fn snapshot_for_tool(tool_name: &str, tool_args: &Value) -> FileSnapshot {
-    if tool_name != "pybun_run" {
-        return FileSnapshot::new();
-    }
-
-    let Some(paths) = tool_args
-        .get("sandbox_policy")
-        .and_then(|p| p.get("allow_write"))
-        .and_then(|v| v.as_array())
-    else {
-        return FileSnapshot::new();
-    };
-
-    let mut snapshot = FileSnapshot::new();
-    for path in paths.iter().filter_map(|v| v.as_str()) {
-        collect_file_snapshot(&normalized_absolute_path(Path::new(path)), &mut snapshot);
-    }
-    snapshot
-}
-
-fn string_array_arg(value: &Value, key: &str) -> Vec<String> {
-    value
-        .get(key)
-        .and_then(|v| v.as_array())
-        .map(|arr| {
-            arr.iter()
-                .filter_map(|v| v.as_str())
-                .filter(|name| !name.is_empty() && !name.contains('=') && !name.contains('\0'))
-                .map(String::from)
-                .collect()
-        })
-        .unwrap_or_default()
-}
-
-fn mcp_sandbox_config_from_policy(policy: &Value) -> crate::sandbox::SandboxConfig {
-    crate::sandbox::SandboxConfig {
-        allow_network: policy
-            .get("allow_network")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false),
-        allow_read: string_array_arg(policy, "allow_read"),
-        allow_write: string_array_arg(policy, "allow_write"),
-        allow_env: string_array_arg(policy, "allow_env")
-            .into_iter()
-            .filter(|name| !crate::sandbox::is_credential_env_name(name))
-            .collect(),
-        timeout_secs: policy
-            .get("timeout_secs")
-            .and_then(|v| v.as_u64())
-            .unwrap_or(crate::sandbox::DEFAULT_SANDBOX_TIMEOUT_SECS),
-        ..Default::default()
-    }
-}
-
-fn sandbox_policy_json(config: &crate::sandbox::SandboxConfig) -> Value {
-    json!({
-        "allow_network": config.allow_network,
-        "allow_read": config.allow_read,
-        "allow_write": config.allow_write,
-        "allow_env": config.allow_env,
-        "timeout_secs": config.timeout_secs,
-        "memory_limit_mb": config.memory_limit_mb,
-        "cpu_limit_secs": config.cpu_limit_secs,
-        "max_processes": config.max_processes,
-        "file_size_limit_mb": config.file_size_limit_mb,
-    })
-}
-
-fn unsafe_no_sandbox_warning(enabled: bool) -> Vec<Value> {
-    if enabled {
-        vec![json!({
-            "level": "warning",
-            "code": "W_MCP_UNSAFE_NO_SANDBOX",
-            "message": "pybun_run executed without the default MCP sandbox because unsafe_no_sandbox=true was set",
-            "suggestion": "Only set unsafe_no_sandbox=true in controlled environments."
-        })]
-    } else {
-        vec![]
-    }
-}
-
-fn describe_run_target(script: Option<&str>, code: Option<&str>) -> String {
-    if let Some(script) = script {
-        format!("Python script: {script}")
-    } else if let Some(code) = code {
-        format!("Python inline code ({} bytes)", code.len())
-    } else {
-        "No Python target".to_string()
-    }
-}
-
-fn estimate_run_risk(
-    script: Option<&str>,
-    code: Option<&str>,
-    sandboxed: bool,
-) -> (String, Vec<String>) {
-    let mut reasons = Vec::new();
-    let source = code.or(script).unwrap_or_default();
-    for (needle, reason) in [
-        (
-            "shutil.rmtree",
-            "recursive file deletion detected: shutil.rmtree",
-        ),
-        ("os.remove", "file deletion detected: os.remove"),
-        ("os.unlink", "file deletion detected: os.unlink"),
-        ("subprocess", "subprocess execution detected"),
-        ("socket", "network access detected"),
-        ("ctypes", "native library access detected: ctypes"),
-        ("cffi", "native library access detected: cffi"),
-    ] {
-        if source.contains(needle) {
-            reasons.push(reason.to_string());
-        }
-    }
-    if !sandboxed {
-        reasons.push("sandbox disabled by unsafe_no_sandbox=true".to_string());
-    }
-
-    let risk = if reasons
-        .iter()
-        .any(|reason| reason.contains("deletion") || reason.contains("sandbox disabled"))
-    {
-        "high"
-    } else if reasons.is_empty() {
-        "low"
-    } else {
-        "medium"
-    };
-
-    (risk.to_string(), reasons)
-}
-
-fn collect_file_snapshot(path: &Path, snapshot: &mut FileSnapshot) {
-    let Ok(metadata) = fs::metadata(path) else {
-        return;
-    };
-    if metadata.is_file() {
-        if let Ok(sha256) = crate::security::sha256_file(path) {
-            snapshot.insert(
-                normalized_absolute_path(path),
-                FileState {
-                    size_bytes: metadata.len(),
-                    sha256,
-                },
-            );
-        }
-        return;
-    }
-
-    if metadata.is_dir()
-        && let Ok(entries) = fs::read_dir(path)
-    {
-        for entry in entries.flatten() {
-            collect_file_snapshot(&entry.path(), snapshot);
-        }
-    }
-}
-
-fn diff_file_writes(before: &FileSnapshot, after: &FileSnapshot) -> Vec<AuditFileWrite> {
-    after
-        .iter()
-        .filter_map(|(path, after_state)| {
-            let before_state = before.get(path);
-            let changed = before_state
-                .map(|state| {
-                    state.size_bytes != after_state.size_bytes || state.sha256 != after_state.sha256
-                })
-                .unwrap_or(true);
-            changed.then(|| AuditFileWrite {
-                path: path.display().to_string(),
-                size_bytes: after_state.size_bytes,
-                sha256_before: before_state.map(|state| state.sha256.clone()),
-                sha256_after: Some(after_state.sha256.clone()),
-            })
-        })
-        .collect()
-}
-
-impl JsonRpcResponse {
-    pub fn success(id: Value, result: Value) -> Self {
-        Self {
-            jsonrpc: "2.0".to_string(),
-            result: Some(result),
-            error: None,
-            id,
-        }
-    }
-
-    pub fn error(id: Value, code: i32, message: impl Into<String>) -> Self {
-        Self {
-            jsonrpc: "2.0".to_string(),
-            result: None,
-            error: Some(JsonRpcError {
-                code,
-                message: message.into(),
-                data: None,
-            }),
-            id,
-        }
-    }
-}
-
-/// MCP Server state
-pub struct McpServer {
-    initialized: bool,
-    session_id: String,
-    audit_log: McpAuditLog,
-}
-
-impl McpServer {
-    pub fn new() -> Self {
-        Self {
-            initialized: false,
-            session_id: Uuid::new_v4().to_string(),
-            audit_log: McpAuditLog::new(),
-        }
-    }
-
-    /// Handle a JSON-RPC request
-    pub async fn handle_request(&mut self, request: JsonRpcRequest) -> Option<JsonRpcResponse> {
-        // Check for notifications that we explicitly handle
-        match request.method.as_str() {
-            "initialized" | "notifications/initialized" => {
-                return None;
-            }
-            _ => {}
-        }
-
-        // For all other methods, if there is no ID, it is a notification and we must not respond
-        let id = match request.id {
-            Some(id) => id,
-            None => return None,
-        };
-
-        match request.method.as_str() {
-            "initialize" => Some(self.handle_initialize(id, request.params)),
-            "tools/list" => Some(self.handle_tools_list(id)),
-            "tools/call" => Some(self.handle_tools_call(id, request.params).await),
-            "resources/list" => Some(self.handle_resources_list(id)),
-            "resources/read" => Some(self.handle_resources_read(id, request.params)),
-            "shutdown" => {
-                eprintln!("MCP server shutting down");
-                Some(JsonRpcResponse::success(id, json!({})))
-            }
-            _ => Some(JsonRpcResponse::error(
-                id,
-                -32601,
-                format!("Method not found: {}", request.method),
-            )),
-        }
-    }
-
-    fn handle_initialize(&mut self, id: Value, _params: Value) -> JsonRpcResponse {
-        self.initialized = true;
-
-        JsonRpcResponse::success(
-            id,
-            json!({
-                "protocolVersion": PROTOCOL_VERSION,
-                "capabilities": {
-                    "tools": {},
-                    "resources": {}
-                },
-                "serverInfo": {
-                    "name": SERVER_NAME,
-                    "version": SERVER_VERSION
-                }
-            }),
-        )
-    }
-
-    fn handle_tools_list(&self, id: Value) -> JsonRpcResponse {
-        let tools = vec![
-            Tool {
-                name: "pybun_resolve".to_string(),
-                description: "Resolve Python package dependencies".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "requirements": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "List of requirements (e.g., ['requests>=2.28', 'flask'])"
-                        },
-                        "pre": {
-                            "type": "boolean",
-                            "description": "Allow pre-release and dev versions when resolving (PEP 440 excludes them by default unless a specifier mentions one; default: false)"
-                        }
-                    },
-                    "required": ["requirements"]
-                }),
-            },
-            Tool {
-                name: "pybun_install".to_string(),
-                description: "Install Python packages. Delegates to the same real install path as CLI `pybun install`: resolves dependencies (falling back to PyPI when no `index` is given), verifies real sha256 hashes, downloads wheels, and installs them into the target environment. Reports \"installed\" only when wheels were actually downloaded and installed; otherwise reports an honest \"resolved\" status.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "requirements": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "List of requirements to install. If omitted, dependencies are read from pyproject.toml."
-                        },
-                        "offline": {
-                            "type": "boolean",
-                            "description": "Use offline mode (cache only); honored by the underlying install path"
-                        },
-                        "system": {
-                            "type": "boolean",
-                            "description": "Allow installing into system Python instead of creating a project-local .pybun/venv (default: false)"
-                        },
-                        "pre": {
-                            "type": "boolean",
-                            "description": "Allow pre-release and dev versions when resolving (PEP 440 excludes them by default unless a specifier mentions one; default: false)"
-                        },
-                        "index": {
-                            "type": "string",
-                            "description": "Path to a local index JSON file. If omitted, falls back to PyPI (same as the CLI)."
-                        },
-                        "lock": {
-                            "type": "string",
-                            "description": "Path to write the lockfile (default: pybun.lockb)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_run".to_string(),
-                description: "Run a Python script inside the default MCP sandbox unless explicitly opted out.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "script": {
-                            "type": "string",
-                            "description": "Path to the Python script"
-                        },
-                        "code": {
-                            "type": "string",
-                            "description": "Inline Python code to execute"
-                        },
-                        "args": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "Arguments to pass to the script"
-                        },
-                        "dry_run": {
-                            "type": "boolean",
-                            "description": "Return an execution plan and risk summary without running code"
-                        },
-                        "unsafe_no_sandbox": {
-                            "type": "boolean",
-                            "description": "Disable the default MCP sandbox. Use only in controlled environments."
-                        },
-                        "sandbox_policy": {
-                            "type": "object",
-                            "description": "Optional sandbox policy. MCP pybun_run is sandboxed by default; use unsafe_no_sandbox only for explicit opt-out.",
-                            "properties": {
-                                "allow_network": {
-                                    "type": "boolean",
-                                    "description": "Allow network access (default: false)"
-                                },
-                                "allow_read": {
-                                    "type": "array",
-                                    "items": {"type": "string"},
-                                    "description": "Paths allowed for reading (empty = no restriction)"
-                                },
-                                "allow_write": {
-                                    "type": "array",
-                                    "items": {"type": "string"},
-                                    "description": "Paths allowed for writing (empty = no restriction)"
-                                },
-                                "allow_env": {
-                                    "type": "array",
-                                    "items": {"type": "string"},
-                                    "description": "Additional non-secret environment variable names to pass through"
-                                },
-                                "timeout_secs": {
-                                    "type": "integer",
-                                    "description": "Maximum wall-clock execution time in seconds"
-                                }
-                            }
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_gc".to_string(),
-                description: "Run garbage collection on PyBun cache".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "max_size": {
-                            "type": "string",
-                            "description": "Maximum cache size (e.g., '1G', '500M')"
-                        },
-                        "dry_run": {
-                            "type": "boolean",
-                            "description": "Preview without deleting"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_doctor".to_string(),
-                description: "Run environment diagnostics".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "verbose": {
-                            "type": "boolean",
-                            "description": "Include verbose diagnostics"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_lint".to_string(),
-                description: "Run linting on Python code and return structured violations. Uses ruff if available.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "script": {
-                            "type": "string",
-                            "description": "Path to Python file or directory to lint"
-                        },
-                        "code": {
-                            "type": "string",
-                            "description": "Inline Python code to lint (written to a temp file)"
-                        },
-                        "select": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "Rule codes to enable (e.g. ['E501', 'F401'])"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_type_check".to_string(),
-                description: "Run type checking on Python code using mypy. Returns structured type errors with hints.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "script": {
-                            "type": "string",
-                            "description": "Path to Python file or directory to type-check"
-                        },
-                        "code": {
-                            "type": "string",
-                            "description": "Inline Python code to type-check"
-                        },
-                        "strict": {
-                            "type": "boolean",
-                            "description": "Enable strict mypy mode"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_profile".to_string(),
-                description: "Profile a Python script using cProfile and return performance hotspots.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "script": {
-                            "type": "string",
-                            "description": "Path to Python script to profile"
-                        },
-                        "code": {
-                            "type": "string",
-                            "description": "Inline Python code to profile"
-                        },
-                        "top_n": {
-                            "type": "integer",
-                            "description": "Number of top hotspots to return (default: 10)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_fix".to_string(),
-                description: "Auto-fix lint violations in a Python file using ruff. Returns a summary of applied fixes.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "script": {
-                            "type": "string",
-                            "description": "Path to Python file to fix (required)"
-                        },
-                        "select": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "Rule codes to fix (default: all auto-fixable)"
-                        },
-                        "unsafe_fixes": {
-                            "type": "boolean",
-                            "description": "Apply unsafe fixes (default: false)"
-                        }
-                    },
-                    "required": ["script"]
-                }),
-            },
-            Tool {
-                name: "pybun_context".to_string(),
-                description: "Return a single-call project state snapshot for agent consumption. Aggregates Python version, venv status, lockfile status, installed packages, and doctor warnings into one structured response.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "summary_only": {
-                            "type": "boolean",
-                            "description": "Return package counts instead of the full installed_packages list (reduces output size for low-token contexts)"
-                        },
-                        "include_drift": {
-                            "type": "boolean",
-                            "description": "Include AST-based import drift analysis (reserved for future use; currently returns null)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_test".to_string(),
-                description: "Run Python tests and return structured per-test results. Returns summary, failures (with rerun_command), and passed entries for agent-friendly test workflows.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "path": {
-                            "type": "string",
-                            "description": "Test file or directory to run (default: auto-discover from current directory)"
-                        },
-                        "changed": {
-                            "type": "boolean",
-                            "description": "Run only tests in files changed since last git commit (new untracked test files are also included)"
-                        },
-                        "fail_fast": {
-                            "type": "boolean",
-                            "description": "Stop on first failure"
-                        },
-                        "filter": {
-                            "type": "string",
-                            "description": "Test name pattern filter (runs only tests whose name contains this string)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_drift".to_string(),
-                description: "Detect dependency drift: undeclared imports (packages imported but not in pyproject.toml) and unused declarations (packages in pyproject.toml but never imported). Returns structured results with agent-callable next_action for pybun_add/pybun_remove.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "cwd": {
-                            "type": "string",
-                            "description": "Directory to analyze (defaults to current working directory)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_audit".to_string(),
-                description: "Scan installed packages for known vulnerabilities using the OSV database. Returns structured results with severity levels and agent-callable fix suggestions via next_action.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "fix": {
-                            "type": "boolean",
-                            "description": "Populate next_action with the pybun_upgrade call needed to fix each vulnerability (default: true)"
-                        },
-                        "severity_threshold": {
-                            "type": "string",
-                            "enum": ["low", "medium", "high", "critical"],
-                            "description": "Only report vulnerabilities at or above this severity level (default: low)"
-                        }
-                    }
-                }),
-            },
-            Tool {
-                name: "pybun_upgrade".to_string(),
-                description: "Upgrade a single package to a specific version in the current environment. Intended to be called by agents acting on next_action entries from pybun_audit.".to_string(),
-                input_schema: json!({
-                    "type": "object",
-                    "properties": {
-                        "package": {
-                            "type": "string",
-                            "description": "Package name to upgrade"
-                        },
-                        "version": {
-                            "type": "string",
-                            "description": "Target version (e.g. '2.31.0')"
-                        }
-                    },
-                    "required": ["package", "version"]
-                }),
-            },
-        ];
-
-        JsonRpcResponse::success(id, json!({ "tools": tools }))
-    }
-
-    async fn handle_tools_call(&mut self, id: Value, params: Value) -> JsonRpcResponse {
-        let tool_name = params.get("name").and_then(|n| n.as_str()).unwrap_or("");
-        let tool_args = params.get("arguments").cloned().unwrap_or(json!({}));
-
-        let start = Instant::now();
-        let before_snapshot = snapshot_for_tool(tool_name, &tool_args);
-        let result = if let Err(err) = self.audit_log.prepare_for_call(&tool_args) {
-            Err(err)
-        } else {
-            match tool_name {
-                "pybun_resolve" => self.call_resolve(tool_args.clone()).await,
-                "pybun_install" => self.call_install(tool_args.clone()).await,
-                "pybun_run" => self.call_run(tool_args.clone()).await,
-                "pybun_gc" => self.call_gc(tool_args.clone()),
-                "pybun_doctor" => self.call_doctor(tool_args.clone()),
-                "pybun_lint" => self.call_lint(tool_args.clone()),
-                "pybun_type_check" => self.call_type_check(tool_args.clone()),
-                "pybun_profile" => self.call_profile(tool_args.clone()),
-                "pybun_fix" => self.call_fix(tool_args.clone()),
-                "pybun_context" => self.call_context(tool_args.clone()),
-                "pybun_drift" => self.call_drift(tool_args.clone()),
-                "pybun_test" => self.call_test(tool_args.clone()),
-                "pybun_audit" => self.call_audit(tool_args.clone()).await,
-                "pybun_upgrade" => self.call_upgrade(tool_args.clone()),
-                _ => Err(format!("Unknown tool: {}", tool_name)),
-            }
-        };
-        let after_snapshot = snapshot_for_tool(tool_name, &tool_args);
-        let file_writes = diff_file_writes(&before_snapshot, &after_snapshot);
-        let duration_ms = start.elapsed().as_millis().min(u128::from(u64::MAX)) as u64;
-
-        let audit_entry = build_audit_entry(
-            &self.session_id,
-            tool_name,
-            &tool_args,
-            &result,
-            self.audit_log.config.hash_inputs,
-            file_writes,
-            duration_ms,
-        );
-        self.audit_log.record(audit_entry);
-
-        match result {
-            Ok(content) => JsonRpcResponse::success(
-                id,
-                json!({
-                    "content": [{
-                        "type": "text",
-                        "text": content
-                    }]
-                }),
-            ),
-            Err(e) => JsonRpcResponse::success(
-                id,
-                json!({
-                    "content": [{
-                        "type": "text",
-                        "text": format!("Error: {}", e)
-                    }],
-                    "isError": true
-                }),
-            ),
-        }
-    }
-
-    fn handle_resources_list(&self, id: Value) -> JsonRpcResponse {
-        let resources = vec![
-            Resource {
-                uri: "pybun://cache/info".to_string(),
-                name: "Cache Information".to_string(),
-                description: Some("Information about the PyBun cache".to_string()),
-                mime_type: Some("application/json".to_string()),
-            },
-            Resource {
-                uri: "pybun://env/info".to_string(),
-                name: "Environment Information".to_string(),
-                description: Some("Current Python environment info".to_string()),
-                mime_type: Some("application/json".to_string()),
-            },
-            Resource {
-                uri: "pybun://audit/recent".to_string(),
-                name: "Recent MCP Audit Entries".to_string(),
-                description: Some("Last 20 audited tool calls for this MCP session".to_string()),
-                mime_type: Some("application/json".to_string()),
-            },
-            Resource {
-                uri: "pybun://project/snapshot".to_string(),
-                name: "Project State Snapshot".to_string(),
-                description: Some("Single-call project state: Python version, venv, lockfile, installed packages, and doctor warnings".to_string()),
-                mime_type: Some("application/json".to_string()),
-            },
-        ];
-
-        JsonRpcResponse::success(id, json!({ "resources": resources }))
-    }
-
-    fn handle_resources_read(&self, id: Value, params: Value) -> JsonRpcResponse {
-        let uri = params.get("uri").and_then(|u| u.as_str()).unwrap_or("");
-
-        let content = match uri {
-            "pybun://cache/info" => self.read_cache_info(),
-            "pybun://env/info" => self.read_env_info(),
-            "pybun://audit/recent" => Ok(self.read_audit_recent()),
-            "pybun://project/snapshot" => self.call_context(json!({})),
-            _ => Err(format!("Unknown resource: {}", uri)),
-        };
-
-        match content {
-            Ok(text) => JsonRpcResponse::success(
-                id,
-                json!({
-                    "contents": [{
-                        "uri": uri,
-                        "mimeType": "application/json",
-                        "text": text
-                    }]
-                }),
-            ),
-            Err(e) => JsonRpcResponse::error(id, -32602, e),
-        }
-    }
-
-    fn read_audit_recent(&self) -> String {
-        self.audit_log.recent_json(&self.session_id)
-    }
-
-    // Tool implementations
-    async fn call_resolve(&self, args: Value) -> Result<String, String> {
-        use crate::index::load_index_from_path;
-        use crate::resolver::{Requirement, ResolveOptions, resolve_with_options};
-
-        let requirements: Vec<String> = args
-            .get("requirements")
-            .and_then(|r| r.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        if requirements.is_empty() {
-            return Err("No requirements provided".to_string());
-        }
-
-        // Parse requirements
-        let parsed_reqs: Vec<Requirement> = requirements
-            .iter()
-            .map(|s| s.parse().unwrap_or_else(|_| Requirement::any(s.trim())))
-            .collect();
-
-        // Opt-in to pre-release versions (mirrors the CLI `--pre` flag).
-        let resolve_options = ResolveOptions {
-            allow_prerelease: args.get("pre").and_then(|p| p.as_bool()).unwrap_or(false),
-            python_version: crate::commands::resolve_target_python_version(),
-        };
-
-        // Try to load index from common locations
-        let index_path = args
-            .get("index")
-            .and_then(|i| i.as_str())
-            .map(PathBuf::from);
-
-        // If index path provided, use it; otherwise try default locations
-        let index_result: Result<_, String> = if let Some(path) = index_path {
-            load_index_from_path(&path).map_err(|e| e.to_string())
-        } else {
-            // Try fixtures/index.json for testing, then fail gracefully
-            let default_paths = vec![
-                PathBuf::from("fixtures/index.json"),
-                PathBuf::from("tests/fixtures/index.json"),
-            ];
-            let mut result: Result<_, String> = Err("No index file found".to_string());
-            for path in default_paths {
-                if path.exists() {
-                    result = load_index_from_path(&path).map_err(|e| e.to_string());
-                    if result.is_ok() {
-                        break;
-                    }
-                }
-            }
-            result
-        };
-
-        match index_result {
-            Ok(index) => {
-                match resolve_with_options(parsed_reqs.clone(), &index, resolve_options).await {
-                    Ok(resolution) => {
-                        let packages: Vec<Value> = resolution
-                        .packages
-                        .values()
-                        .map(|pkg| {
-                            json!({
-                                "name": pkg.name,
-                                "version": pkg.version,
-                                "dependencies": pkg.dependencies.iter().map(|d| d.to_string()).collect::<Vec<_>>(),
-                            })
-                        })
-                        .collect();
-
-                        let prerelease_fallbacks: Vec<Value> = resolution
-                            .prerelease_fallbacks
-                            .iter()
-                            .map(|p| json!({ "name": p.name, "version": p.version }))
-                            .collect();
-
-                        Ok(json!({
-                            "status": "resolved",
-                            "requirements": requirements,
-                            "packages": packages,
-                            "count": resolution.packages.len(),
-                            "prerelease_fallbacks": prerelease_fallbacks,
-                        })
-                        .to_string())
-                    }
-                    Err(e) => Err(format!("Resolution failed: {}", e)),
-                }
-            }
-            Err(e) => {
-                // Return a partial result indicating index is not available
-                Ok(json!({
-                    "status": "no_index",
-                    "requirements": requirements,
-                    "message": format!("Could not load package index: {}. Provide 'index' path in arguments.", e),
-                    "parsed_requirements": parsed_reqs.iter().map(|r| r.to_string()).collect::<Vec<_>>(),
-                })
-                .to_string())
-            }
-        }
-    }
-
-    /// Install packages by delegating to the same `install()` implementation
-    /// used by the CLI `pybun install` command (`crate::commands::install`).
-    ///
-    /// This intentionally reuses the CLI's real index selection (with PyPI
-    /// fallback when no fixture/local `index` is given), real hash
-    /// verification (no `sha256:placeholder`), real wheel download, and real
-    /// wheel installation into the target environment's site-packages.
-    ///
-    /// Issue #284: previously this method never downloaded or installed any
-    /// wheel, yet unconditionally reported `"status": "installed"` /
-    /// `"Resolved and installed N packages"`. It now reports `"installed"`
-    /// only when `InstallOutcome::installed_count > 0` (i.e. wheels were
-    /// actually fetched and installed); otherwise it reports an honest
-    /// `"resolved"` status describing exactly what happened (dependency
-    /// resolution and lockfile generation, without any wheel installation).
-    async fn call_install(&self, args: Value) -> Result<String, String> {
-        use crate::resolver::Requirement;
-        use crate::schema::EventCollector;
-
-        let requirements: Vec<String> = args
-            .get("requirements")
-            .and_then(|r| r.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let parsed_requirements: Vec<Requirement> = requirements
-            .iter()
-            .map(|s| s.parse().unwrap_or_else(|_| Requirement::any(s.trim())))
-            .collect();
-
-        // Honor `offline` (previously accepted but silently ignored).
-        let offline = args
-            .get("offline")
-            .and_then(|o| o.as_bool())
-            .unwrap_or(false);
-
-        // Honor an explicit opt-in to installing into system Python, matching
-        // the CLI's `--system` safety guard (defaults to creating an isolated
-        // `.pybun/venv` rather than touching system Python).
-        let system = args
-            .get("system")
-            .and_then(|s| s.as_bool())
-            .unwrap_or(false);
-
-        let index = args
-            .get("index")
-            .and_then(|i| i.as_str())
-            .map(PathBuf::from);
-
-        // Default lockfile name matches the CLI project path (`pybun.lockb`),
-        // not the previous MCP-only `pybun.lock` (see PR-A3 for the broader
-        // MCP/CLI naming-unification track; this fix inherits the CLI
-        // default as a side effect of delegating to the real install path).
-        let lock = args
-            .get("lock")
-            .and_then(|l| l.as_str())
-            .map(PathBuf::from)
-            .unwrap_or_else(|| PathBuf::from("pybun.lockb"));
-
-        // Opt-in to pre-release versions (mirrors the CLI `--pre` flag).
-        let pre = args.get("pre").and_then(|p| p.as_bool()).unwrap_or(false);
-
-        let install_args = crate::cli::InstallArgs {
-            offline,
-            system,
-            requirements: parsed_requirements,
-            index,
-            lock,
-            workspace: false,
-            member: None,
-            group: None,
-            pre,
-        };
-
-        let mut collector = EventCollector::new();
-        let result = crate::commands::install(&install_args, &mut collector).await;
-        let diagnostics = serde_json::to_value(collector.into_diagnostics()).unwrap_or(Value::Null);
-
-        match result {
-            Ok(outcome) => {
-                let really_installed = outcome.installed_count > 0;
-                let status = if really_installed {
-                    "installed"
-                } else {
-                    "resolved"
-                };
-                let message = if really_installed {
-                    format!(
-                        "Installed {} wheel(s) for {} resolved package(s) -> {}",
-                        outcome.installed_count,
-                        outcome.packages.len(),
-                        outcome.lockfile.display()
-                    )
-                } else {
-                    format!(
-                        "Resolved {} package(s) and wrote {} (no wheels were downloaded or installed)",
-                        outcome.packages.len(),
-                        outcome.lockfile.display()
-                    )
-                };
-
-                Ok(json!({
-                    "status": status,
-                    "packages": outcome.packages,
-                    "lockfile": outcome.lockfile.display().to_string(),
-                    "count": outcome.packages.len(),
-                    "installed_count": outcome.installed_count,
-                    "verified": outcome.verified,
-                    "artifacts": outcome.artifacts,
-                    "message": message,
-                    "diagnostics": diagnostics,
-                })
-                .to_string())
-            }
-            Err(e) => Err(e.to_string()),
-        }
-    }
-
-    async fn call_run(&self, args: Value) -> Result<String, String> {
-        use crate::env::find_python_env;
-        use crate::schema::EventCollector;
-
-        let script = args.get("script").and_then(|s| s.as_str());
-        let code = args.get("code").and_then(|c| c.as_str());
-        let run_args: Vec<String> = args
-            .get("args")
-            .and_then(|a| a.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let unsafe_no_sandbox = args
-            .get("unsafe_no_sandbox")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let use_sandbox = !unsafe_no_sandbox;
-        let policy = args
-            .get("sandbox_policy")
-            .cloned()
-            .unwrap_or_else(|| json!({}));
-        let effective_sandbox_config = mcp_sandbox_config_from_policy(&policy);
-        let warnings = unsafe_no_sandbox_warning(unsafe_no_sandbox);
-
-        let dry_run = args
-            .get("dry_run")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        if script.is_none() && code.is_none() {
-            return Err("Either 'script' or 'code' must be provided".to_string());
-        }
-        if let Some(script_path) = script
-            && !PathBuf::from(script_path).exists()
-        {
-            return Err(format!("Script not found: {}", script_path));
-        }
-        if dry_run {
-            let (risk_level, risk_reasons) = estimate_run_risk(script, code, use_sandbox);
-            return Ok(json!({
-                "status": "dry_run",
-                "dry_run": true,
-                "would_execute": describe_run_target(script, code),
-                "sandboxed": use_sandbox,
-                "sandbox_policy": use_sandbox.then(|| sandbox_policy_json(&effective_sandbox_config)),
-                "risk_level": risk_level,
-                "risk_reasons": risk_reasons,
-                "warnings": warnings,
-            })
-            .to_string());
-        }
-
-        // Delegate to the same `commands::run_script` implementation the CLI's
-        // `pybun run` uses, so PEP 723 dependency auto-install, sandboxing, and
-        // interpreter discovery cannot silently diverge between the MCP and CLI
-        // entry points (Issue #272). Interpreter discovery and sandboxed execution
-        // were already shared via `env::find_python_env` /
-        // `sandbox::execute_with_optional_sandbox`; this closes the remaining gap
-        // where MCP skipped PEP 723 dependency install entirely.
-        let run_args_struct = crate::cli::RunArgs {
-            target: script.map(|s| s.to_string()),
-            code: code.map(|s| s.to_string()),
-            sandbox: use_sandbox,
-            allow_network: effective_sandbox_config.allow_network,
-            allow_read: effective_sandbox_config.allow_read.clone(),
-            allow_write: effective_sandbox_config.allow_write.clone(),
-            allow_env: effective_sandbox_config.allow_env.clone(),
-            sandbox_timeout: effective_sandbox_config.timeout_secs,
-            sandbox_memory: effective_sandbox_config.memory_limit_mb,
-            sandbox_cpu: effective_sandbox_config.cpu_limit_secs,
-            profile: "dev".to_string(),
-            passthrough: run_args,
-        };
-
-        // Reported for informational purposes only; `run_script` performs its own
-        // (identical) interpreter discovery internally.
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-        let python_path = find_python_env(&working_dir)
-            .map(|env| env.python_path.to_string_lossy().to_string())
-            .unwrap_or_default();
-
-        let mut collector = EventCollector::new();
-        let result = crate::commands::run_script(
-            &run_args_struct,
-            &mut collector,
-            crate::cli::OutputFormat::Json,
-        )
-        .await;
-
-        match result {
-            Ok(outcome) => {
-                // Enrich diagnostics with a structured traceback when the script
-                // failed, mirroring the CLI `pybun run` dispatcher (Issue #266).
-                if outcome.exit_code != 0
-                    && let Some(tb) = outcome.stderr.as_deref().and_then(crate::traceback::parse)
-                {
-                    let mut diag = crate::schema::Diagnostic::error(tb.message.clone());
-                    diag.code = Some(tb.code);
-                    diag.file = tb.location.as_ref().map(|l| l.file.clone());
-                    diag.line = tb.location.as_ref().map(|l| l.line);
-                    diag.exception_type = Some(tb.exception_type);
-                    diag.location = tb.location.as_ref().map(|loc| {
-                        json!({
-                            "file": loc.file,
-                            "line": loc.line,
-                            "function": loc.function,
-                        })
-                    });
-                    diag.next_action = tb.next_action.map(|a| {
-                        json!({
-                            "tool": a.tool,
-                            "args": a.args,
-                        })
-                    });
-                    collector.diagnostic(diag);
-                }
-                let diagnostics =
-                    serde_json::to_value(collector.into_diagnostics()).unwrap_or(Value::Null);
-
-                let sandboxed = outcome.sandbox.as_ref().map(|s| s.enabled).unwrap_or(false);
-                let audit = outcome.sandbox.as_ref().and_then(|s| s.audit.clone());
-                let resource_limits = outcome.sandbox.as_ref().map(|s| s.resource_limits.clone());
-                let timed_out = outcome
-                    .sandbox
-                    .as_ref()
-                    .map(|s| s.timed_out)
-                    .unwrap_or(false);
-
-                Ok(json!({
-                    "status": if outcome.exit_code == 0 { "success" } else { "error" },
-                    "target": outcome.target,
-                    "exit_code": outcome.exit_code,
-                    "stdout": outcome.stdout.unwrap_or_default(),
-                    "stderr": outcome.stderr.unwrap_or_default(),
-                    "python": python_path,
-                    "sandboxed": sandboxed,
-                    "audit": audit,
-                    "resource_limits": resource_limits,
-                    "timed_out": timed_out,
-                    "diagnostics": diagnostics,
-                    "warnings": warnings,
-                    "pep723_dependencies": outcome.pep723_deps,
-                    "pep723_backend": outcome.pep723_backend,
-                    "temp_env": outcome.temp_env,
-                    "cleanup": outcome.cleanup,
-                    "cache_hit": outcome.cache_hit,
-                })
-                .to_string())
-            }
-            Err(e) => Err(e.to_string()),
-        }
-    }
-
-    fn call_gc(&self, args: Value) -> Result<String, String> {
-        let max_size = args.get("max_size").and_then(|s| s.as_str());
-        let dry_run = args
-            .get("dry_run")
-            .and_then(|d| d.as_bool())
-            .unwrap_or(false);
-
-        use crate::cache::{Cache, format_size, parse_size};
-
-        let cache = Cache::new().map_err(|e| e.to_string())?;
-        let max_bytes = max_size.map(parse_size).transpose()?;
-
-        let result = cache.gc(max_bytes, dry_run).map_err(|e| e.to_string())?;
-
-        Ok(json!({
-            "status": "gc_complete",
-            "freed_bytes": result.freed_bytes,
-            "freed_human": format_size(result.freed_bytes),
-            "files_removed": result.files_removed,
-            "dry_run": dry_run
-        })
-        .to_string())
-    }
-
-    fn call_doctor(&self, args: Value) -> Result<String, String> {
-        use crate::cache::Cache;
-        use crate::env::find_python_env;
-        use crate::project::Project;
-
-        let verbose = args
-            .get("verbose")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let mut checks: Vec<Value> = Vec::new();
-        let mut all_ok = true;
-
-        // Check Python availability
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-        match find_python_env(&working_dir) {
-            Ok(env) => {
-                checks.push(json!({
-                    "name": "python",
-                    "status": "ok",
-                    "message": format!("Python found at {}", env.python_path.display()),
-                    "source": format!("{}", env.source),
-                    "version": env.version,
-                }));
-            }
-            Err(e) => {
-                checks.push(json!({
-                    "name": "python",
-                    "status": "error",
-                    "message": format!("Python not found: {}", e),
-                }));
-                all_ok = false;
-            }
-        }
-
-        // Check cache directory
-        match Cache::new() {
-            Ok(cache) => {
-                let cache_dir = cache.root();
-                let mut cache_check = json!({
-                    "name": "cache",
-                    "status": "ok",
-                    "message": format!("Cache directory: {}", cache_dir.display()),
-                    "path": cache_dir.display().to_string(),
-                });
-
-                if verbose && let Ok(size) = cache.total_size() {
-                    cache_check["total_size"] = json!(size);
-                    cache_check["total_size_human"] = json!(crate::cache::format_size(size));
-                }
-                checks.push(cache_check);
-            }
-            Err(e) => {
-                checks.push(json!({
-                    "name": "cache",
-                    "status": "error",
-                    "message": format!("Cache initialization failed: {}", e),
-                }));
-                all_ok = false;
-            }
-        }
-
-        // Check the PyPI metadata cache directory (separate from the main
-        // cache root above - see issue #202). Flag stale/corrupt entries
-        // (e.g. unreadable `.bin`/`.json` files left over from an older or
-        // interrupted `pybun` run, see issue #268) as a non-fatal `info`
-        // status pointing at `pybun gc` as the fix, mirroring the CLI
-        // `pybun doctor` check and the `corrupt` lockfile status below.
-        if let Some(pypi_cache_dir) = crate::pypi::pypi_cache_dir() {
-            let stats = crate::pypi::pypi_cache_stats(&pypi_cache_dir);
-            let status = if stats.stale_count > 0 { "info" } else { "ok" };
-            checks.push(json!({
-                "name": "pypi_cache",
-                "status": status,
-                "message": format!(
-                    "PyPI metadata cache: {} ({} entries, {} stale)",
-                    pypi_cache_dir.display(),
-                    stats.entry_count,
-                    stats.stale_count,
-                ),
-                "path": pypi_cache_dir.display().to_string(),
-                "entry_count": stats.entry_count,
-                "total_bytes": stats.total_bytes,
-                "stale_count": stats.stale_count,
-                "hint": if stats.stale_count > 0 {
-                    json!("Run `pybun gc` to remove corrupt/stale PyPI cache entries")
-                } else {
-                    Value::Null
-                },
-            }));
-            // Stale/corrupt cache entries are non-fatal (self-healed on
-            // next fetch) - they do not flip `all_ok` to false.
-        }
-
-        // Check for pyproject.toml
-        match Project::discover(&working_dir) {
-            Ok(project) => {
-                let deps = project.dependencies();
-                checks.push(json!({
-                    "name": "project",
-                    "status": "ok",
-                    "message": format!("Project found at {}", project.path().display()),
-                    "path": project.path().display().to_string(),
-                    "dependencies_count": deps.len(),
-                    "dependencies": if verbose { json!(deps) } else { json!(null) },
-                }));
-            }
-            Err(_) => {
-                checks.push(json!({
-                    "name": "project",
-                    "status": "info",
-                    "message": "No pyproject.toml found in current directory",
-                }));
-            }
-        }
-
-        // Check for lockfile
-        let lockfile_path = working_dir.join("pybun.lock");
-        if lockfile_path.exists() {
-            checks.push(json!({
-                "name": "lockfile",
-                "status": "ok",
-                "message": format!("Lockfile found at {}", lockfile_path.display()),
-                "path": lockfile_path.display().to_string(),
-            }));
-        } else {
-            checks.push(json!({
-                "name": "lockfile",
-                "status": "info",
-                "message": "No pybun.lock found",
-            }));
-        }
-
-        let status = if all_ok { "healthy" } else { "issues_found" };
-        let summary = if all_ok {
-            "All checks passed"
-        } else {
-            "Some issues found"
-        };
-
-        Ok(json!({
-            "status": status,
-            "checks": checks,
-            "verbose": verbose,
-            "message": summary,
-        })
-        .to_string())
-    }
-
-    fn call_lint(&self, args: Value) -> Result<String, String> {
-        use crate::env::find_python_env;
-
-        let script = args.get("script").and_then(|s| s.as_str());
-        let code = args.get("code").and_then(|c| c.as_str());
-        let select: Vec<String> = args
-            .get("select")
-            .and_then(|s| s.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        // Determine target path (script or temp file for inline code)
-        let (target_path, _temp_file) = match (script, code) {
-            (Some(p), _) => (p.to_string(), None::<tempfile::NamedTempFile>),
-            (None, Some(inline)) => {
-                let mut tmp = tempfile::Builder::new()
-                    .suffix(".py")
-                    .tempfile()
-                    .map_err(|e| format!("Failed to create temp file: {}", e))?;
-                use std::io::Write;
-                write!(tmp, "{}", inline).map_err(|e| e.to_string())?;
-                let path = tmp.path().to_string_lossy().to_string();
-                (path, Some(tmp))
-            }
-            _ => return Err("Either 'script' or 'code' must be provided".to_string()),
-        };
-
-        let target_display = if script.is_none() {
-            "inline_code"
-        } else {
-            &target_path
-        };
-
-        let working_dir = current_working_dir()?;
-
-        if let Some(ruff_runner) = resolve_ruff_runner(&working_dir)? {
-            let mut cmd = ruff_runner.command();
-            cmd.args(["check", "--output-format=json"]);
-            if !select.is_empty() {
-                cmd.arg("--select");
-                cmd.arg(select.join(","));
-            }
-            cmd.arg(&target_path);
-
-            let output = cmd
-                .output()
-                .map_err(|e| format!("Failed to run ruff: {}", e))?;
-            let stdout_str = String::from_utf8_lossy(&output.stdout).to_string();
-            let stderr_str = String::from_utf8_lossy(&output.stderr).to_string();
-
-            if !valid_ruff_status(&output.status) {
-                return Err(ruff_failure_message("check", &stdout_str, &stderr_str));
-            }
-
-            let violations: Vec<Value> = parse_ruff_violations(&stdout_str)?
-                .into_iter()
-                .map(|v| {
-                    json!({
-                        "file": v.get("filename").and_then(|f| f.as_str()).unwrap_or(&target_path),
-                        "line": v.get("location").and_then(|l| l.get("row")).and_then(|r| r.as_u64()).unwrap_or(0),
-                        "column": v.get("location").and_then(|l| l.get("column")).and_then(|c| c.as_u64()).unwrap_or(0),
-                        "code": v.get("code").and_then(|c| c.as_str()).unwrap_or(""),
-                        "message": v.get("message").and_then(|m| m.as_str()).unwrap_or(""),
-                        "fix_available": v.get("fix").is_some(),
-                    })
-                })
-                .collect();
-
-            Ok(json!({
-                "status": "lint_complete",
-                "tool": "ruff",
-                "target": target_display,
-                "violations": violations,
-                "violation_count": violations.len(),
-                "clean": violations.is_empty(),
-                "diagnostics": violations.iter().filter_map(|v| {
-                    let msg = v.get("message")?.as_str()?;
-                    let code = v.get("code")?.as_str()?;
-                    Some(json!({
-                        "kind": code,
-                        "message": msg,
-                        "hint": if v.get("fix_available").and_then(|f| f.as_bool()).unwrap_or(false) {
-                            format!("Auto-fixable with pybun_fix. Code: {}", code)
-                        } else {
-                            format!("Manual fix required. Code: {}", code)
-                        }
-                    }))
-                }).collect::<Vec<_>>(),
-            })
-            .to_string())
-        } else {
-            // Fall back to python -m py_compile for basic syntax check
-            let env = find_python_env(&working_dir).map_err(|e| e.to_string())?;
-            let python_path = env.python_path.to_string_lossy().to_string();
-
-            let output = ProcessCommand::new(&python_path)
-                .args(["-m", "py_compile", &target_path])
-                .output()
-                .map_err(|e| format!("Failed to run py_compile: {}", e))?;
-
-            let stderr = String::from_utf8_lossy(&output.stderr).to_string();
-            Ok(json!({
-                "status": "lint_complete",
-                "tool": "py_compile",
-                "tool_not_available": "ruff",
-                "hint": "Install ruff for full linting: pybun add ruff",
-                "violations": [],
-                "syntax_ok": output.status.success(),
-                "stderr": stderr,
-                "target": target_display,
-            })
-            .to_string())
-        }
-    }
-
-    fn call_type_check(&self, args: Value) -> Result<String, String> {
-        use crate::env::find_python_env;
-
-        let script = args.get("script").and_then(|s| s.as_str());
-        let code = args.get("code").and_then(|c| c.as_str());
-        let strict = args
-            .get("strict")
-            .and_then(|s| s.as_bool())
-            .unwrap_or(false);
-
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-        let env = find_python_env(&working_dir).map_err(|e| e.to_string())?;
-        let python_path = env.python_path.to_string_lossy().to_string();
-
-        // Determine target (script or temp file for inline code)
-        let (target_path, _temp_file) = match (script, code) {
-            (Some(p), _) => (p.to_string(), None::<tempfile::NamedTempFile>),
-            (None, Some(inline)) => {
-                let mut tmp = tempfile::Builder::new()
-                    .suffix(".py")
-                    .tempfile()
-                    .map_err(|e| format!("Failed to create temp file: {}", e))?;
-                use std::io::Write;
-                write!(tmp, "{}", inline).map_err(|e| e.to_string())?;
-                let path = tmp.path().to_string_lossy().to_string();
-                (path, Some(tmp))
-            }
-            _ => return Err("Either 'script' or 'code' must be provided".to_string()),
-        };
-
-        // Check if mypy is available
-        let mypy_check = ProcessCommand::new(&python_path)
-            .args(["-m", "mypy", "--version"])
-            .output();
-
-        if mypy_check.is_err() || !mypy_check.unwrap().status.success() {
-            return Ok(json!({
-                "status": "type_check_complete",
-                "tool_not_available": "mypy",
-                "hint": "Install mypy for type checking: pybun add mypy",
-                "errors": [],
-                "target": target_path,
-            })
-            .to_string());
-        }
-
-        // Run mypy
-        let mut cmd = ProcessCommand::new(&python_path);
-        cmd.args(["-m", "mypy", "--show-error-codes", "--no-color-output"]);
-        if strict {
-            cmd.arg("--strict");
-        }
-        cmd.arg(&target_path);
-
-        let output = cmd
-            .output()
-            .map_err(|e| format!("Failed to run mypy: {}", e))?;
-
-        let stdout_str = String::from_utf8_lossy(&output.stdout).to_string();
-
-        // Parse mypy output (line format: "file:line: severity: message  [code]")
-        let errors: Vec<Value> = stdout_str
-            .lines()
-            .filter(|line| line.contains(": error:") || line.contains(": warning:") || line.contains(": note:"))
-            .filter_map(|line| {
-                // Parse: path:line:col: severity: message  [error-code]
-                let parts: Vec<&str> = line.splitn(4, ':').collect();
-                if parts.len() < 4 {
-                    return None;
-                }
-                let file = parts[0];
-                let line_num: u64 = parts[1].trim().parse().unwrap_or(0);
-                let rest = parts[3];
-                let (severity, message) = if let Some(idx) = rest.find(": ") {
-                    let sev = rest[..idx].trim();
-                    let msg = rest[idx + 2..].trim();
-                    (sev, msg)
-                } else {
-                    ("error", rest.trim())
-                };
-                // Extract error code if present: "message  [error-code]"
-                let (msg_clean, error_code) = if let (Some(start), Some(end)) = (message.rfind('['), message.rfind(']')) {
-                    let code = &message[start + 1..end];
-                    let msg = message[..start].trim();
-                    (msg, code.to_string())
-                } else {
-                    (message, String::new())
-                };
-
-                Some(json!({
-                    "file": file,
-                    "line": line_num,
-                    "severity": severity.trim(),
-                    "message": msg_clean,
-                    "code": error_code,
-                    "hint": format!("See https://mypy.readthedocs.io/en/stable/error_codes.html#{}", error_code.to_lowercase()),
-                }))
-            })
-            .collect();
-
-        let target_display = if script.is_none() {
-            "inline_code"
-        } else {
-            &target_path
-        };
-
-        Ok(json!({
-            "status": "type_check_complete",
-            "tool": "mypy",
-            "target": target_display,
-            "strict": strict,
-            "success": output.status.success(),
-            "errors": errors,
-            "error_count": errors.len(),
-            "clean": errors.is_empty(),
-            "raw_output": stdout_str,
-        })
-        .to_string())
-    }
-
-    fn call_profile(&self, args: Value) -> Result<String, String> {
-        use crate::env::find_python_env;
-
-        let script = args.get("script").and_then(|s| s.as_str());
-        let code = args.get("code").and_then(|c| c.as_str());
-        let top_n = args.get("top_n").and_then(|n| n.as_u64()).unwrap_or(10) as usize;
-
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-        let env = find_python_env(&working_dir).map_err(|e| e.to_string())?;
-        let python_path = env.python_path.to_string_lossy().to_string();
-
-        // Resolve target: write inline code to temp file if needed
-        let (_temp_target, target_path_str): (Option<tempfile::NamedTempFile>, String) =
-            match (script, code) {
-                (Some(p), _) => {
-                    let path = PathBuf::from(p);
-                    if !path.exists() {
-                        return Err(format!("Script not found: {}", p));
-                    }
-                    (None, p.to_string())
-                }
-                (None, Some(inline)) => {
-                    let mut tmp = tempfile::Builder::new()
-                        .suffix(".py")
-                        .tempfile()
-                        .map_err(|e| format!("Failed to create temp file: {}", e))?;
-                    use std::io::Write as _;
-                    write!(tmp, "{}", inline).map_err(|e| e.to_string())?;
-                    let p = tmp.path().to_string_lossy().to_string();
-                    (Some(tmp), p)
-                }
-                _ => return Err("Either 'script' or 'code' must be provided".to_string()),
-            };
-
-        // Write profiler runner to a temp file to avoid format-string escaping issues
-        // with Python dict literals inside Rust format! macros.
-        let profiler_src = [
-            "import cProfile, pstats, io, json, os, re, sys",
-            &format!(
-                "_target = {}",
-                serde_json::to_string(&target_path_str).unwrap_or_default()
-            ),
-            &format!("_top_n = {}", top_n),
-            "_globals = {'__name__': '__main__', '__file__': _target, '__package__': None, '__cached__': None}",
-            "_argv = list(sys.argv)",
-            "_path = list(sys.path)",
-            "_script_dir = os.path.dirname(_target)",
-            "if _script_dir:",
-            "    sys.path.insert(0, _script_dir)",
-            "sys.argv = [_target]",
-            "pr = cProfile.Profile()",
-            "try:",
-            "    pr.enable()",
-            "    with open(_target) as _f:",
-            "        exec(compile(_f.read(), _target, 'exec'), _globals)",
-            "finally:",
-            "    pr.disable()",
-            "    sys.argv = _argv",
-            "    sys.path[:] = _path",
-            "s = io.StringIO()",
-            "ps = pstats.Stats(pr, stream=s).sort_stats('cumulative')",
-            "ps.print_stats(_top_n)",
-            "raw = s.getvalue()",
-            "hotspots = []",
-            "for line in raw.strip().split('\\n'):",
-            "    parts = line.split()",
-            "    if len(parts) >= 6 and parts[0].replace('.','',1).isdigit():",
-            "        try:",
-            "            hotspots.append({'ncalls': parts[0], 'tottime': float(parts[1]),",
-            "                'percall_tot': float(parts[2]), 'cumtime': float(parts[3]),",
-            "                'percall_cum': float(parts[4]), 'location': ' '.join(parts[5:])})",
-            "        except (ValueError, IndexError):",
-            "            pass",
-            "total_match = re.search(r'(\\d+\\.\\d+) seconds', raw)",
-            "total_time = float(total_match.group(1)) if total_match else 0.0",
-            "calls_match = re.search(r'(\\d+) function calls', raw)",
-            "total_calls = int(calls_match.group(1)) if calls_match else 0",
-            "print(json.dumps({'total_time_s': total_time, 'total_calls': total_calls,",
-            "    'hotspots': hotspots[:_top_n], 'status': 'profile_complete'}))",
-        ]
-        .join("\n");
-
-        let mut runner_tmp = tempfile::Builder::new()
-            .suffix(".py")
-            .tempfile()
-            .map_err(|e| format!("Failed to create profiler script: {}", e))?;
-        {
-            use std::io::Write as _;
-            write!(runner_tmp, "{}", profiler_src).map_err(|e| e.to_string())?;
-        }
-
-        let profile_script_path = runner_tmp.path().to_string_lossy().to_string();
-
-        let output = ProcessCommand::new(&python_path)
-            .arg(&profile_script_path)
-            .output()
-            .map_err(|e| format!("Failed to run profiler: {}", e))?;
-
-        let stdout_str = String::from_utf8_lossy(&output.stdout).to_string();
-        let stderr_str = String::from_utf8_lossy(&output.stderr).to_string();
-
-        if !output.status.success() {
-            return Err(format!(
-                "Profiler failed: {}",
-                if stderr_str.is_empty() {
-                    &stdout_str
-                } else {
-                    &stderr_str
-                }
-            ));
-        }
-
-        // Parse the JSON output from the profiler script
-        let result: Value = serde_json::from_str(stdout_str.trim()).unwrap_or_else(|_| {
-            json!({
-                "status": "profile_complete",
-                "raw": stdout_str,
-            })
-        });
-
-        let target_display = script.unwrap_or("inline_code");
-        let mut out = result.as_object().cloned().unwrap_or_default();
-        out.insert("target".to_string(), json!(target_display));
-        out.insert("top_n".to_string(), json!(top_n));
-
-        serde_json::to_string(&out).map_err(|e| e.to_string())
-    }
-
-    fn call_fix(&self, args: Value) -> Result<String, String> {
-        let script = args
-            .get("script")
-            .and_then(|s| s.as_str())
-            .ok_or_else(|| "'script' is required for pybun_fix".to_string())?;
-
-        let select: Vec<String> = args
-            .get("select")
-            .and_then(|s| s.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default();
-
-        let unsafe_fixes = args
-            .get("unsafe_fixes")
-            .and_then(|u| u.as_bool())
-            .unwrap_or(false);
-
-        let path = PathBuf::from(script);
-        if !path.exists() {
-            return Err(format!("Script not found: {}", script));
-        }
-
-        let working_dir = current_working_dir()?;
-        let Some(ruff_runner) = resolve_ruff_runner(&working_dir)? else {
-            return Ok(json!({
-                "status": "fix_complete",
-                "tool_not_available": "ruff",
-                "hint": "Install ruff for auto-fixing: pybun add ruff",
-                "fixes_applied": 0,
-                "target": script,
-            })
-            .to_string());
-        };
-
-        // Run ruff check --fix to get count of fixable violations before
-        let mut check_cmd = ruff_runner.command();
-        check_cmd.args(["check", "--output-format=json"]);
-        if !select.is_empty() {
-            check_cmd.arg("--select");
-            check_cmd.arg(select.join(","));
-        }
-        check_cmd.arg(script);
-        let before_output = check_cmd.output().map_err(|e| e.to_string())?;
-        let before_str = String::from_utf8_lossy(&before_output.stdout).to_string();
-        let before_stderr = String::from_utf8_lossy(&before_output.stderr).to_string();
-        if !valid_ruff_status(&before_output.status) {
-            return Err(ruff_failure_message("check", &before_str, &before_stderr));
-        }
-        let before_count = parse_ruff_violations(&before_str)?.len();
-
-        // Apply fixes
-        let mut fix_cmd = ruff_runner.command();
-        fix_cmd.args(["check", "--fix"]);
-        if unsafe_fixes {
-            fix_cmd.arg("--unsafe-fixes");
-        }
-        if !select.is_empty() {
-            fix_cmd.arg("--select");
-            fix_cmd.arg(select.join(","));
-        }
-        fix_cmd.arg(script);
-
-        let fix_output = fix_cmd
-            .output()
-            .map_err(|e| format!("Failed to run ruff fix: {}", e))?;
-        let fix_stdout = String::from_utf8_lossy(&fix_output.stdout).to_string();
-        let fix_stderr = String::from_utf8_lossy(&fix_output.stderr).to_string();
-        if !valid_ruff_status(&fix_output.status) {
-            return Err(ruff_failure_message(
-                "check --fix",
-                &fix_stdout,
-                &fix_stderr,
-            ));
-        }
-
-        // Count remaining violations
-        let mut recheck_cmd = ruff_runner.command();
-        recheck_cmd.args(["check", "--output-format=json"]);
-        if !select.is_empty() {
-            recheck_cmd.arg("--select");
-            recheck_cmd.arg(select.join(","));
-        }
-        recheck_cmd.arg(script);
-        let after_output = recheck_cmd.output().map_err(|e| e.to_string())?;
-        let after_str = String::from_utf8_lossy(&after_output.stdout).to_string();
-        let after_stderr = String::from_utf8_lossy(&after_output.stderr).to_string();
-        if !valid_ruff_status(&after_output.status) {
-            return Err(ruff_failure_message("check", &after_str, &after_stderr));
-        }
-        let after_count = parse_ruff_violations(&after_str)?.len();
-
-        let fixes_applied = before_count.saturating_sub(after_count);
-
-        Ok(json!({
-            "status": "fix_complete",
-            "tool": "ruff",
-            "target": script,
-            "fixes_applied": fixes_applied,
-            "violations_before": before_count,
-            "violations_after": after_count,
-            "unsafe_fixes": unsafe_fixes,
-            "stderr": fix_stderr,
-        })
-        .to_string())
-    }
-
-    // Resource implementations
-    fn read_cache_info(&self) -> Result<String, String> {
-        use crate::cache::{Cache, format_size};
-
-        let cache = Cache::new().map_err(|e| e.to_string())?;
-        let total_size = cache.total_size().map_err(|e| e.to_string())?;
-
-        Ok(json!({
-            "root": cache.root().display().to_string(),
-            "total_size": total_size,
-            "total_size_human": format_size(total_size)
-        })
-        .to_string())
-    }
-
-    fn read_env_info(&self) -> Result<String, String> {
-        use crate::env::find_python_env;
-
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-
-        match find_python_env(&working_dir) {
-            Ok(env) => Ok(json!({
-                "python_path": env.python_path.display().to_string(),
-                "source": format!("{}", env.source),
-                "version": env.version
-            })
-            .to_string()),
-            Err(e) => Ok(json!({
-                "error": e.to_string(),
-                "message": "No Python environment found"
-            })
-            .to_string()),
-        }
-    }
-
-    fn call_context(&self, args: Value) -> Result<String, String> {
-        use crate::env::{EnvSource, find_python_env};
-        use crate::lockfile::Lockfile;
-        use crate::project::Project;
-        use std::collections::HashSet;
-        use std::time::{SystemTime, UNIX_EPOCH};
-
-        let summary_only = args
-            .get("summary_only")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let include_drift = args
-            .get("include_drift")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-
-        let working_dir = if let Some(cwd) = args.get("cwd").and_then(|v| v.as_str()) {
-            std::path::PathBuf::from(cwd)
-        } else {
-            std::env::current_dir().map_err(|e| e.to_string())?
-        };
-
-        let snapshot_at_ms = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_millis() as u64)
-            .unwrap_or(0);
-
-        // PEP 503: normalize [-_.] sequences to a single hyphen for package name comparison.
-        fn normalize_pkg_name(name: &str) -> String {
-            let lower = name.to_lowercase();
-            let mut result = String::with_capacity(lower.len());
-            let mut prev_was_sep = false;
-            for ch in lower.chars() {
-                if ch == '-' || ch == '_' || ch == '.' {
-                    if !prev_was_sep {
-                        result.push('-');
-                    }
-                    prev_was_sep = true;
-                } else {
-                    result.push(ch);
-                    prev_was_sep = false;
-                }
-            }
-            result
-        }
-
-        // ── Python / venv info ────────────────────────────────────────────────
-        let (python_version, venv_path, venv_status) = match find_python_env(&working_dir) {
-            Ok(env) => {
-                let is_local = matches!(env.source, EnvSource::ProjectLocal | EnvSource::PybunEnv);
-                let venv = if is_local {
-                    env.python_path
-                        .parent()
-                        .and_then(|p| p.parent())
-                        .map(|p| p.display().to_string())
-                } else {
-                    None
-                };
-                (env.version, venv, "ok".to_string())
-            }
-            Err(_) => {
-                // Mirror find_project_venv's probe list: directory exists but no Python binary = corrupt.
-                let venv_candidates = [
-                    working_dir.join(".pybun").join("venv"),
-                    working_dir.join(".venv"),
-                    working_dir.join("venv"),
-                ];
-                let status = if venv_candidates.iter().any(|p| p.is_dir()) {
-                    "corrupt"
-                } else {
-                    "missing"
-                };
-                (None, None, status.to_string())
-            }
-        };
-
-        // ── Project declared dependencies ─────────────────────────────────────
-        let declared_deps: Vec<String> = Project::discover(&working_dir)
-            .map(|p| p.dependencies())
-            .unwrap_or_default();
-        let declared_names: HashSet<String> = declared_deps
-            .iter()
-            .map(|d| {
-                let bare = d
-                    .split(['>', '<', '=', '!', '~', ';', '['])
-                    .next()
-                    .unwrap_or(d)
-                    .trim();
-                normalize_pkg_name(bare)
-            })
-            .collect();
-
-        // ── Lockfile ──────────────────────────────────────────────────────────
-        let lockfile_path = working_dir.join("pybun.lock");
-        let (lockfile_status, locked_packages, locked_names) =
-            if let Ok(bytes) = std::fs::read(&lockfile_path) {
-                match Lockfile::from_bytes(&bytes) {
-                    Ok(lf) => {
-                        let locked_names: HashSet<String> =
-                            lf.packages.keys().map(|k| normalize_pkg_name(k)).collect();
-                        let locked: Vec<Value> = lf
-                            .packages
-                            .values()
-                            .map(|pkg| {
-                                let norm = normalize_pkg_name(&pkg.name);
-                                let declared = declared_names.contains(&norm);
-                                json!({
-                                    "name": pkg.name,
-                                    "version": pkg.version,
-                                    "declared": declared,
-                                    "locked": true
-                                })
-                            })
-                            .collect();
-                        let has_drift = declared_names
-                            .iter()
-                            .any(|d| !locked_names.contains(d.as_str()));
-                        let status = if has_drift { "drift" } else { "in_sync" };
-                        (status.to_string(), locked, Some(locked_names))
-                    }
-                    // Parse failure = lockfile exists but is unreadable/corrupt, distinct from drift.
-                    Err(_) => ("corrupt".to_string(), vec![], None),
-                }
-            } else {
-                ("missing".to_string(), vec![], None)
-            };
-
-        // ── Doctor warnings ───────────────────────────────────────────────────
-        let mut doctor_warnings: Vec<Value> = Vec::new();
-        if venv_status == "missing" {
-            doctor_warnings.push(json!({
-                "code": "W_VENV_MISSING",
-                "message": "No virtual environment found. Run `pybun install` to create one.",
-                "severity": "warn"
-            }));
-        }
-        if venv_status == "corrupt" {
-            doctor_warnings.push(json!({
-                "code": "W_VENV_CORRUPT",
-                "message": "Virtual environment directory exists but Python binary is missing or invalid.",
-                "severity": "error"
-            }));
-        }
-        if lockfile_status == "missing" {
-            doctor_warnings.push(json!({
-                "code": "W_LOCKFILE_MISSING",
-                "message": "No pybun.lock found. Run `pybun install` to generate one.",
-                "severity": "info"
-            }));
-        }
-        if lockfile_status == "corrupt" {
-            doctor_warnings.push(json!({
-                "code": "W_LOCKFILE_CORRUPT",
-                "message": "pybun.lock exists but cannot be parsed. Run `pybun install` to regenerate.",
-                "severity": "error"
-            }));
-        }
-        if lockfile_status == "drift" {
-            doctor_warnings.push(json!({
-                "code": "W_LOCKFILE_DRIFT",
-                "message": "Declared dependencies differ from lockfile. Run `pybun install` to sync.",
-                "severity": "warn"
-            }));
-        }
-
-        // ── Import drift analysis (only when include_drift=true) ──────────────
-        let import_drift_summary = if include_drift {
-            let drift_result = crate::drift::analyze(&working_dir);
-            json!({
-                "undeclared_imports": drift_result.undeclared_imports,
-                "unused_declarations": drift_result.unused_declarations,
-                "files_scanned": drift_result.files_scanned,
-                "analysis_notes": drift_result.analysis_notes,
-            })
-        } else {
-            json!({
-                "undeclared_imports": null,
-                "outdated_packages": []
-            })
-        };
-
-        // ── Assemble response ─────────────────────────────────────────────────
-        if summary_only {
-            let declared_count = declared_names.len();
-            let installed_count = locked_packages.len();
-            let drift_count = locked_names.as_ref().map_or(0, |names| {
-                declared_names
-                    .iter()
-                    .filter(|d| !names.contains(d.as_str()))
-                    .count()
-            });
-
-            Ok(json!({
-                "python_version": python_version,
-                "venv_path": venv_path,
-                "venv_status": venv_status,
-                "lockfile_status": lockfile_status,
-                "installed_count": installed_count,
-                "declared_count": declared_count,
-                "drift_count": drift_count,
-                "doctor_warnings": doctor_warnings,
-                "drift_summary": import_drift_summary,
-                "snapshot_at_ms": snapshot_at_ms
-            })
-            .to_string())
-        } else {
-            Ok(json!({
-                "python_version": python_version,
-                "venv_path": venv_path,
-                "venv_status": venv_status,
-                "lockfile_status": lockfile_status,
-                "installed_packages": locked_packages,
-                "doctor_warnings": doctor_warnings,
-                "drift_summary": import_drift_summary,
-                "snapshot_at_ms": snapshot_at_ms
-            })
-            .to_string())
-        }
-    }
-
-    fn call_drift(&self, args: Value) -> Result<String, String> {
-        use crate::drift;
-
-        let cwd_path = if let Some(cwd) = args.get("cwd").and_then(|v| v.as_str()) {
-            std::path::PathBuf::from(cwd)
-        } else {
-            std::env::current_dir().map_err(|e| format!("failed to get cwd: {e}"))?
-        };
-
-        if !cwd_path.join("pyproject.toml").exists() {
-            return Ok(serde_json::to_string(&json!({
-                "error": "pyproject.toml not found",
-                "undeclared_imports": [],
-                "unused_declarations": [],
-                "analysis_notes": ["pyproject.toml not found"],
-                "files_scanned": 0
-            }))
-            .unwrap());
-        }
-
-        let result = drift::analyze(&cwd_path);
-        Ok(serde_json::to_string(&json!({
-            "undeclared_imports": result.undeclared_imports,
-            "unused_declarations": result.unused_declarations,
-            "analysis_notes": result.analysis_notes,
-            "files_scanned": result.files_scanned,
-        }))
-        .unwrap())
-    }
-
-    fn call_test(&self, args: Value) -> Result<String, String> {
-        use crate::test_discovery::{TestDiscovery, TestItemType};
-        use crate::test_executor::{ExecutorConfig, TestExecutor, TestOutcome};
-
-        let working_dir = current_working_dir()?;
-
-        let path = args.get("path").and_then(|p| p.as_str()).map(String::from);
-        let changed = args
-            .get("changed")
-            .and_then(|c| c.as_bool())
-            .unwrap_or(false);
-        let fail_fast = args
-            .get("fail_fast")
-            .and_then(|f| f.as_bool())
-            .unwrap_or(false);
-        let filter = args
-            .get("filter")
-            .and_then(|f| f.as_str())
-            .map(String::from);
-
-        // `path` and `changed` are mutually exclusive
-        if changed && path.is_some() {
-            return Err(
-                "'path' and 'changed' are mutually exclusive: use 'changed' to run tests in \
-                 git-modified files, or 'path' to target a specific file or directory"
-                    .to_string(),
-            );
-        }
-
-        let (search_paths, analysis_notes): (Vec<std::path::PathBuf>, Vec<&str>) = if changed {
-            let (paths, git_status) = get_changed_test_files(&working_dir)?;
-            let note = match git_status {
-                GitAvailability::NotARepo => {
-                    return Ok(json!({
-                        "summary": {
-                            "total": 0, "passed": 0, "failed": 0,
-                            "skipped": 0, "errors": 0, "duration_ms": 0
-                        },
-                        "failures": [],
-                        "passed": [],
-                        "analysis_notes": [
-                            "git is not available or this directory is not a git repository; \
-                             cannot determine changed files"
-                        ]
-                    })
-                    .to_string());
-                }
-                GitAvailability::Available if paths.is_empty() => {
-                    return Ok(json!({
-                        "summary": {
-                            "total": 0, "passed": 0, "failed": 0,
-                            "skipped": 0, "errors": 0, "duration_ms": 0
-                        },
-                        "failures": [],
-                        "passed": [],
-                        "analysis_notes": ["No changed test files detected since last commit"]
-                    })
-                    .to_string());
-                }
-                GitAvailability::Available => "running tests from git-changed files only",
-            };
-            (paths, vec![note])
-        } else if let Some(p) = path {
-            // Guard against path traversal: reject paths that escape working_dir
-            let joined = working_dir.join(&p);
-            let canonical = joined
-                .canonicalize()
-                .map_err(|e| format!("path '{}' does not exist or cannot be resolved: {}", p, e))?;
-            if !canonical.starts_with(&working_dir) {
-                return Err(format!(
-                    "path '{}' is outside the working directory and is not allowed",
-                    p
-                ));
-            }
-            (vec![canonical], vec![])
-        } else {
-            (vec![working_dir.clone()], vec![])
-        };
-
-        if search_paths.is_empty() {
-            return Ok(json!({
-                "summary": {
-                    "total": 0, "passed": 0, "failed": 0,
-                    "skipped": 0, "errors": 0, "duration_ms": 0
-                },
-                "failures": [],
-                "passed": [],
-                "analysis_notes": analysis_notes
-            })
-            .to_string());
-        }
-
-        let discovery = TestDiscovery::new();
-        let discovery_result = discovery.discover(&search_paths);
-
-        let mut tests: Vec<crate::test_discovery::TestItem> = discovery_result
-            .tests
-            .iter()
-            .filter(|t| t.item_type != TestItemType::Class)
-            .cloned()
-            .collect();
-
-        if let Some(ref pattern) = filter {
-            tests.retain(|t| {
-                t.name.contains(pattern.as_str()) || t.short_name.contains(pattern.as_str())
-            });
-        }
-
-        if tests.is_empty() {
-            return Ok(json!({
-                "summary": {
-                    "total": 0, "passed": 0, "failed": 0,
-                    "skipped": 0, "errors": 0, "duration_ms": 0
-                },
-                "failures": [],
-                "passed": [],
-            })
-            .to_string());
-        }
-
-        let python = crate::env::find_python_env(&working_dir)
-            .map(|env| env.python_path.to_string_lossy().to_string())
-            .unwrap_or_else(|_| "python3".to_string());
-
-        let workers = std::thread::available_parallelism()
-            .map(|p| p.get())
-            .unwrap_or(4);
-
-        let config = ExecutorConfig {
-            workers,
-            fail_fast,
-            shard: None,
-            verbose: false,
-            timeout: None,
-            retries: 0,
-            python,
-        };
-
-        let executor = TestExecutor::new(config);
-        let result = executor.execute(tests);
-        let summary = &result.summary;
-
-        let failures: Vec<Value> = result
-            .results
-            .iter()
-            .filter(|r| {
-                !matches!(
-                    r.outcome,
-                    TestOutcome::Passed | TestOutcome::Skipped | TestOutcome::XFail
-                )
-            })
-            .map(|r| {
-                let file = r.path.display().to_string();
-                let rerun_command = format!("pybun test {}::{}", file, r.name);
-                let message = if !r.stderr.is_empty() {
-                    r.stderr.clone()
-                } else {
-                    r.stdout.clone()
-                };
-                json!({
-                    "name": r.name,
-                    "file": file,
-                    "line": r.line,
-                    "duration_ms": r.duration_ms,
-                    "status": serde_json::to_value(&r.outcome).unwrap_or(json!("failed")),
-                    "message": message,
-                    "rerun_command": rerun_command,
-                })
-            })
-            .collect();
-
-        let passed: Vec<Value> = result
-            .results
-            .iter()
-            .filter(|r| matches!(r.outcome, TestOutcome::Passed))
-            .map(|r| {
-                json!({
-                    "name": r.name,
-                    "file": r.path.display().to_string(),
-                    "line": r.line,
-                    "duration_ms": r.duration_ms,
-                    "status": "passed",
-                })
-            })
-            .collect();
-
-        Ok(json!({
-            "summary": {
-                "total": summary.total,
-                "passed": summary.passed,
-                "failed": summary.failed,
-                "skipped": summary.skipped,
-                "errors": summary.errors,
-                "duration_ms": summary.duration_ms,
-            },
-            "failures": failures,
-            "passed": passed,
-            "analysis_notes": analysis_notes,
-        })
-        .to_string())
-    }
-
-    async fn call_audit(&self, args: Value) -> Result<String, String> {
-        use crate::audit::{list_installed_packages, scan_for_vulnerabilities};
-        use crate::env::find_python_env;
-
-        let fix = args.get("fix").and_then(|v| v.as_bool()).unwrap_or(true);
-        let severity_threshold = args
-            .get("severity_threshold")
-            .and_then(|s| s.as_str())
-            .unwrap_or("low");
-
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-
-        // Fail open (empty package list) on env/pip errors, consistent with
-        // this tool's pre-existing behavior: an agent calling pybun_audit
-        // should get a best-effort scan, not a hard error, if the
-        // environment can't be inspected.
-        let packages = match find_python_env(&working_dir) {
-            Ok(env) => list_installed_packages(&env.python_path).unwrap_or_default(),
-            Err(_) => vec![],
-        };
-
-        if packages.is_empty() {
-            return Ok(json!({
-                "status": "ok",
-                "summary": {
-                    "scanned": 0,
-                    "vulnerable": 0,
-                    "critical": 0,
-                    "high": 0,
-                    "medium": 0,
-                    "low": 0
-                },
-                "vulnerabilities": [],
-                "scanner": "osv",
-                "scanner_version": "1.0"
-            })
-            .to_string());
-        }
-
-        let osv_url = crate::audit::default_osv_url();
-        let report = scan_for_vulnerabilities(&packages, &osv_url, severity_threshold).await?;
-
-        let vulnerabilities: Vec<Value> = report
-            .vulnerabilities
-            .iter()
-            .map(|v| {
-                let next_action = if fix {
-                    v.fix_version.as_ref().map(|fv| {
-                        json!({
-                            "tool": "pybun_upgrade",
-                            "args": {
-                                "package": v.package,
-                                "version": fv
-                            }
-                        })
-                    })
-                } else {
-                    None
-                };
-
-                json!({
-                    "package": v.package,
-                    "installed_version": v.installed_version,
-                    "vulnerability_id": v.vulnerability_id,
-                    "severity": v.severity,
-                    "description": v.description,
-                    "fix_version": v.fix_version,
-                    "next_action": next_action
-                })
-            })
-            .collect();
-
-        Ok(json!({
-            "status": "ok",
-            "summary": {
-                "scanned": report.scanned,
-                "vulnerable": vulnerabilities.len(),
-                "critical": report.count_at_severity("critical"),
-                "high": report.count_at_severity("high"),
-                "medium": report.count_at_severity("medium"),
-                "low": report.count_at_severity("low"),
-                "unscanned": report.unscanned
-            },
-            "vulnerabilities": vulnerabilities,
-            "scanner": "osv",
-            "scanner_version": "1.0"
-        })
-        .to_string())
-    }
-
-    fn call_upgrade(&self, args: Value) -> Result<String, String> {
-        use crate::env::find_python_env;
-
-        let package = args
-            .get("package")
-            .and_then(|p| p.as_str())
-            .ok_or_else(|| "Missing required argument: package".to_string())?;
-        let version = args
-            .get("version")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| "Missing required argument: version".to_string())?;
-
-        let working_dir = std::env::current_dir().map_err(|e| e.to_string())?;
-        let env = find_python_env(&working_dir).map_err(|e| e.to_string())?;
-        let python_path = env.python_path.to_string_lossy().to_string();
-
-        let spec = format!("{}=={}", package, version);
-        let output = ProcessCommand::new(&python_path)
-            .args(["-m", "pip", "install", "--disable-pip-version-check", &spec])
-            .output()
-            .map_err(|e| format!("Failed to run pip install: {}", e))?;
-
-        if output.status.success() {
-            Ok(json!({
-                "status": "upgraded",
-                "package": package,
-                "version": version,
-                "spec": spec
-            })
-            .to_string())
-        } else {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            Err(format!("pip install {} failed: {}", spec, stderr.trim()))
-        }
-    }
-}
-
-impl Default for McpServer {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Run the MCP server in stdio mode
-pub async fn run_stdio_server() -> Result<(), Box<dyn std::error::Error>> {
-    eprintln!("PyBun MCP server starting (stdio mode)...");
-
-    let stdin = tokio::io::stdin();
-    let mut stdout = tokio::io::stdout();
-    let reader = BufReader::new(stdin);
-    let mut lines = reader.lines();
-
-    let mut server = McpServer::new();
-
-    while let Ok(Some(line)) = lines.next_line().await {
-        if line.trim().is_empty() {
-            continue;
-        }
-
-        let request: JsonRpcRequest = match serde_json::from_str(&line) {
-            Ok(req) => req,
-            Err(e) => {
-                eprintln!("Invalid JSON-RPC request: {}", e);
-                let error_response = JsonRpcResponse::error(Value::Null, -32700, "Parse error");
-                let _ = stdout
-                    .write_all(serde_json::to_string(&error_response)?.as_bytes())
-                    .await;
-                let _ = stdout.write_all(b"\n").await;
-                let _ = stdout.flush().await;
-                continue;
-            }
-        };
-
-        if let Some(response) = server.handle_request(request).await {
-            let response_json = serde_json::to_string(&response)?;
-            stdout.write_all(response_json.as_bytes()).await?;
-            stdout.write_all(b"\n").await?;
-            stdout.flush().await?;
-        }
-    }
-
-    eprintln!("PyBun MCP server stopped.");
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn test_initialize() {
-        let mut server = McpServer::new();
-        let request = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "initialize".to_string(),
-            params: json!({
-                "protocolVersion": "2024-11-05",
-                "capabilities": {},
-                "clientInfo": {"name": "test", "version": "0.1.0"}
-            }),
-            id: Some(json!(1)),
-        };
-
-        let response = server.handle_request(request).await.unwrap();
-        assert!(response.result.is_some());
-        let result = response.result.unwrap();
-        assert!(result.get("protocolVersion").is_some());
-        assert!(result.get("serverInfo").is_some());
-    }
-
-    #[tokio::test]
-    async fn test_tools_list() {
-        let mut server = McpServer::new();
-        let request = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "tools/list".to_string(),
-            params: json!({}),
-            id: Some(json!(2)),
-        };
-
-        let response = server.handle_request(request).await.unwrap();
-        assert!(response.result.is_some());
-        let result = response.result.unwrap();
-        let tools = result.get("tools").unwrap().as_array().unwrap();
-        assert!(!tools.is_empty());
-
-        // Check some expected tools
-        let tool_names: Vec<&str> = tools
-            .iter()
-            .filter_map(|t| t.get("name").and_then(|n| n.as_str()))
-            .collect();
-        assert!(tool_names.contains(&"pybun_resolve"));
-        assert!(tool_names.contains(&"pybun_install"));
-        assert!(tool_names.contains(&"pybun_run"));
-        assert!(tool_names.contains(&"pybun_gc"));
-    }
-
-    #[tokio::test]
-    async fn test_resources_list() {
-        let mut server = McpServer::new();
-        let request = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "resources/list".to_string(),
-            params: json!({}),
-            id: Some(json!(3)),
-        };
-
-        let response = server.handle_request(request).await.unwrap();
-        assert!(response.result.is_some());
-        let result = response.result.unwrap();
-        let resources = result.get("resources").unwrap().as_array().unwrap();
-        assert!(!resources.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_unknown_method() {
-        let mut server = McpServer::new();
-        let request = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "unknown/method".to_string(),
-            params: json!({}),
-            id: Some(json!(4)),
-        };
-
-        let response = server.handle_request(request).await.unwrap();
-        assert!(response.error.is_some());
-        assert_eq!(response.error.unwrap().code, -32601);
-    }
-
-    #[tokio::test]
-    async fn test_tools_call_gc() {
-        let mut server = McpServer::new();
-        let request = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "tools/call".to_string(),
-            params: json!({
-                "name": "pybun_gc",
-                "arguments": {
-                    "dry_run": true
-                }
-            }),
-            id: Some(json!(5)),
-        };
-
-        let response = server.handle_request(request).await.unwrap();
-        assert!(response.result.is_some());
-    }
-
-    #[tokio::test]
-    async fn test_notification_handling() {
-        let mut server = McpServer::new();
-
-        // 1. "initialized" notification (standard) - should return None
-        let req1 = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "initialized".to_string(),
-            params: json!({}),
-            id: None,
-        };
-        assert!(server.handle_request(req1).await.is_none());
-
-        // 2. "notifications/initialized" (custom) - should return None
-        let req2 = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "notifications/initialized".to_string(),
-            params: json!({}),
-            id: None,
-        };
-        assert!(server.handle_request(req2).await.is_none());
-
-        // 3. "tools/list" as notification (missing id) - should return None (spec compliance)
-        let req3 = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "tools/list".to_string(),
-            params: json!({}),
-            id: None,
-        };
-        assert!(server.handle_request(req3).await.is_none());
-
-        // 4. "unknown/method" as notification (missing id) - should return None (no error)
-        let req4 = JsonRpcRequest {
-            jsonrpc: "2.0".to_string(),
-            method: "unknown/method".to_string(),
-            params: json!({}),
-            id: None,
-        };
-        assert!(server.handle_request(req4).await.is_none());
-    }
-}
